% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Selects.r
\name{A.cal}
\alias{A.cal}
\title{Calculate A matrix}
\usage{
A.cal(s, d)
}
\arguments{
\item{s}{indice of sires}

\item{d}{indice of dams}
}
\value{
additive kinship matrix
}
\description{
Build date: Nov 14, 2018
Last update: Jul 31, 2019
}
\examples{
s <- c(0, 0, 0, 0, 1, 3, 3, 1, 5, 7, 5, 7, 1, 3, 5, 7)
d <- c(0, 0, 0, 0, 2, 4, 4, 2, 6, 8, 8, 6, 6, 8, 4, 8)
A.mat <- A.cal(s, d)
A.mat
}
\author{
Dong Yin
}
