% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduces.r
\name{adj.dam}
\alias{adj.dam}
\title{Adjust ped.dam for correcting litter size}
\usage{
adj.dam(ped.dam, ind.stay, ratio)
}
\arguments{
\item{ped.dam}{indice of dams in the pedigree}

\item{ind.stay}{selected individuals regarded as parents}

\item{ratio}{ratio of males in all individuals}
}
\value{
indice of dams after adjustment
}
\description{
Build date: Nov 14, 2018
Last update: Aug 1, 2019
}
\examples{
basepop <- getpop(nind = 100, from = 1, ratio = 0.5)
ped.dam <- basepop$dam
ind.stay <- basepop$index
idx.dam <- adj.dam(ped.dam = ped.dam, ind.stay = ind.stay, ratio = 0.5)
str(idx.dam)
}
\author{
Dong Yin
}
