% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Genotype.r
\name{annotation}
\alias{annotation}
\title{Get map with annotation}
\usage{
annotation(SP, verbose = TRUE)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{verbose}{whether to print detail}
}
\value{
a map of marker information
}
\description{
Build date: Nov 14, 2018
Last update: Mar 2, 2022
}
\examples{
SP <- param.annot(qtn.num = 10,
                  qtn.model = "A + D + A:D")
SP <- annotation(SP)
head(SP$map$pop.map.GxG)
}
\author{
Dong Yin
}
