% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{build.CV}
\alias{build.CV}
\title{Build design matrix according to covariate}
\usage{
build.CV(cv.name, data)
}
\arguments{
\item{cv.name}{covariate names}

\item{data}{a data frame}
}
\value{
design matrix
}
\description{
Build date: Jan 20, 2020
Last update: Jan 20, 2020
}
\examples{
dat <- data.frame(
f1 = sample(LETTERS[1:3], 20, TRUE),
f2 = sample(LETTERS[4:5], 20, TRUE),
row.names = paste0("id_", 1:20))
cv.name <- c("f1", "f2")
CV <- build.CV(cv.name = cv.name, data = dat)
}
\author{
Dong Yin
}
