% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{build.effMat}
\alias{build.effMat}
\title{Build design matrix according to fixed effects, covariants or random effects}
\usage{
build.effMat(fixed_eff = NULL, covar = NULL, hasMu = TRUE, data)
}
\arguments{
\item{fixed_eff}{the names of fixed effects by vector or list}

\item{covar}{the names of covariants by vector or list}

\item{hasMu}{whether to add mu to effect matrix}

\item{data}{a data frame}
}
\value{
design matrix
}
\description{
Build date: Jan 20, 2020
Last update: Jan 20, 2020
}
\examples{
dat <- data.frame(
f1 = sample(LETTERS[1:3], 20, TRUE),
f2 = sample(LETTERS[4:5], 20, TRUE),
cov = sample.int(20),
row.names = paste0("id_", 1:20))
fixed_eff <- list(c("f1", "f2"), c("f1"))
covar <- list("cov", NULL)
CV <- build.effMat(fixed_eff=fixed_eff, covar=covar, data = dat)
}
\author{
Dong Yin
}
