% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.r
\name{cal.A}
\alias{cal.A}
\title{Calculate for phenotype of A model}
\usage{
cal.A(qtn, h2, eff1, ind.a = NULL, verbose = FALSE)
}
\arguments{
\item{qtn}{QTN matrix of population}

\item{h2}{heritability of trait}

\item{eff1}{effects list of QTN}

\item{ind.a}{additive effects of QTN}

\item{verbose}{whether to show all information}
}
\value{
phenotype of A model
}
\description{
Build date: Nov 14, 2018
Last update: Jul 30, 2019
}
\examples{
num.qtn <- c(2, 6, 10) # three qtn groups
nqtn <- sum(num.qtn) # total number of qtn
nind <- 100
qtn <- matrix(sample(c(0, 1, 2), nqtn*nind, replace=TRUE), nqtn, nind)
eff.a <- cal.eff(num.qtn = num.qtn, eff.var = c(0.4, 0.2, 0.02),
                 dist.qtn = "normal", eff.unit = 0.5, shape = 1, scale = 1)
eff1 <- list(eff.a = eff.a)
h2 <- c(0.3, 0.1, 0.05, 0.05, 0.05, 0.01)
pheno.a1 <- cal.A(qtn = qtn, h2 = h2, eff1 = eff1, ind.a = NULL, verbose = TRUE)
str(pheno.a1)

ind.a <- crossprod(qtn, eff.a)
pheno.a2 <- cal.A(qtn = NULL, h2 = h2, eff1 = NULL, ind.a = ind.a, verbose = TRUE)
str(pheno.a2)
}
\author{
Dong Yin
}
