% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.r
\name{cal.AD}
\alias{cal.AD}
\title{Calculate for phenotype of AD model}
\usage{
cal.AD(qtn, h2, eff1, sel.on = TRUE, inner.env = NULL,
  verbose = FALSE)
}
\arguments{
\item{qtn}{QTN matrix of population}

\item{h2}{heritability of trait}

\item{eff1}{effects list of QTN}

\item{sel.on}{whether to add selection}

\item{inner.env}{environment of main function of simer}

\item{verbose}{whether to show all information}
}
\value{
phenotype of AD model
}
\description{
Build date: Nov 14, 2018
Last update: Jul 30, 2019
}
\examples{
num.qtn <- c(2, 6, 10) # three qtn groups
len.qtn <- length(num.qtn)
nqtn <- sum(num.qtn) # total number of qtn
nind <- 100
qtn <- matrix(sample(c(0, 1, 2), nqtn*nind, replace=TRUE), nqtn, nind)
var.tr1 <- c(0.4, 0.2, 0.02, 0.02, 0.02, 0.02, 0.02, 0.001)
eff.a <- cal.eff(num.qtn = c(2, 6, 10), eff.var = var.tr1[1:len.qtn],
                 dist.qtn = "normal", eff.unit = 0.5, shape = 1, scale = 1)
eff.d <- cal.eff(num.qtn = nqtn,        eff.var = var.tr1[len.qtn+1],
                 dist.qtn = "normal", eff.unit = 0.5, shape = 1, scale = 1)
eff1 <- list(eff.a = eff.a, eff.d = eff.d)
h2 <- c(0.3, 0.1, 0.05, 0.05, 0.05, 0.01)
pheno.ad <- cal.AD(qtn = qtn, h2 = h2, eff1 = eff1, sel.on = TRUE, inner.env = NULL, verbose = TRUE)
str(pheno.ad)
}
\author{
Dong Yin
}
