% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.r
\name{cal.blk}
\alias{cal.blk}
\title{Calculate for block ranges in map}
\usage{
cal.blk(pos.map)
}
\arguments{
\item{pos.map}{map with block information and recombination information}
}
\value{
block ranges
}
\description{
Build date: Aug 15, 2019
Last update: Aug 15, 2019
}
\examples{
# get map with block and recombination information
data(simdata)
nmrk <- nrow(input.map)
pos.map <- check.map(input.map = input.map, num.marker = nmrk, len.block = 5e7)

# calculate for block ranges
blk.rg <- cal.blk(pos.map)
dim(blk.rg)
head(blk.rg)
}
\author{
Dong Yin
}
