% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Phenotype.r
\name{cal.FR}
\alias{cal.FR}
\title{Calculate for fixed effects and random effects}
\usage{
cal.FR(pop = NULL, FR, var.pheno = NULL, pop.env = NULL, verbose = TRUE)
}
\arguments{
\item{pop}{population information}

\item{FR}{list of fixed effects, random effects, and their combination}

\item{var.pheno}{phenotype variances of all traits}

\item{pop.env}{R environment of population information}

\item{verbose}{whether to print detail}
}
\value{
list of fixed effects and random effects
}
\description{
Build date: Nov 1, 2019
Last update: Nov 1, 2019
}
\examples{
pop <- getpop(100, 1, 0.5)
a <- simer.sample(paste0("a", 1:3), 100)
b <- simer.sample(paste0("b", 1:50), 100)
pop$a <- a # load your fixed  effects
pop$b <- b # load your random effects
pop.env <- environment()

# combination of fixed effects
cmb.fix <- list(tr1 = c("mu", "gen", "sex", "a"), # trait 1
                tr2 = c("mu", "diet", "season")) # trait 2
            
fix <- list( 
         mu = list(level = "mu", eff = 20),  				      
        gen = FALSE,         
        sex = list(level = c("1", "2"), eff = c(50, 30)), 
       diet = list(level = c("d1", "d2", "d3"), eff = c(10, 20, 30)),
     season = list(level = c("s1", "s2", "s3", "s4"), eff = c(10, 20, 30, 20)), 
          a = list(level = c("a1", "a2", "a3"), eff = c(10, 20, 30))) 
          
# combination and variance ratio of random effects
tr1 <- list(rn = c("PE"), ratio = 0.1)
tr2 <- list(rn = c("litter", "b"), ratio = c(0.1, 0.1))          
cmb.rand <- list(tr1 = tr1, tr2 = tr2)   
      
rand <- list(       
         PE = list(level = paste0("p", 1:50), eff = rnorm(50)), 
     litter = list(level = paste0("l", 1:50), eff = rnorm(50)),    
          b = list(level = paste0("b", 1:50), eff = rnorm(50)))      
  
FR <- list(cmb.fix = cmb.fix, fix = fix, cmb.rand = cmb.rand, rand = rand)
fr <- cal.FR(pop = pop, FR = FR, var.pheno = c(100, 100), pop.env = pop.env, verbose = TRUE)
str(fr) 	          	         	          	         
}
\author{
Dong Yin
}
