% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.r
\name{cal.genoloc}
\alias{cal.genoloc}
\title{Calculate the location of the individual ID in the genotype matrix}
\usage{
cal.genoloc(stay.id, geno.id)
}
\arguments{
\item{stay.id}{the individual ID}

\item{geno.id}{genotype ID}
}
\value{
the location of the individual ID in the genotype matrix
}
\description{
Build date: May 4, 2020
Last update: May 4, 2020
}
\examples{
stay.id <- c(1, 4, 5)
geno.id <- c(1, 5, 9, 4, 2)
loc <- cal.genoloc(stay.id, geno.id)
loc
}
\author{
Dong Yin
}
