% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.r
\name{cal.gnt}
\alias{cal.gnt}
\title{Generate genetic effects}
\usage{
cal.gnt(
  geno = NULL,
  h2 = NULL,
  effs = NULL,
  sel.on = TRUE,
  inner.env = NULL,
  verbose = TRUE
)
}
\arguments{
\item{geno}{genotype matrix of the population, an individual has two columns}

\item{h2}{heritability vector of the trait, every elements are corresponding to a, d, aXa, aXd, dXa, dXd respectively}

\item{effs}{a list of selected markers and their effects}

\item{sel.on}{whether to add selection}

\item{inner.env}{R environment of parameter "effs"}

\item{verbose}{whether to print detail}
}
\value{
phenotype of population
}
\description{
Build date: Nov 3, 2019
Last update: Nov 3, 2019
}
\examples{
basepop <- getpop(nind = 100, from = 1, ratio = 0.1)
basepop.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)
geno <- geno.cvt1(basepop.geno)
effs <-
    cal.effs(pop.geno = basepop.geno,
             cal.model = "A",
             num.qtn.tr1 = c(2, 6, 10),
             sd.tr1 = c(0.4, 0.2, 0.02, 0.02, 0.02, 0.02, 0.02, 0.001),
             dist.qtn.tr1 = rep("normal", 6),
             prob.tr1 = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5),
             shape.tr1 = c(1, 1, 1, 1, 1, 1),
             scale.tr1 = c(1, 1, 1, 1, 1, 1),
             multrait = FALSE,
             num.qtn.trn = matrix(c(18, 10, 10, 20), 2, 2),
             sd.trn = diag(c(1, 0.5)),
             qtn.spot = rep(0.1, 10),
             maf = 0, 
             verbose = TRUE)
info.eff <- cal.gnt(geno = geno, h2 = c(0.3, 0.1, 0.05, 0.05, 0.05, 0.01), 
    effs = effs, sel.on = TRUE, inner.env = NULL, verbose = TRUE)
str(info.eff)
}
\references{
Kao C and Zeng Z (2002) <https://www.genetics.org/content/160/3/1243.long>
}
\author{
Dong Yin
}
