% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selects.r
\name{cal.idx}
\alias{cal.idx}
\title{Calculate the weight of index selection}
\usage{
cal.idx(pop.pheno, index.wt)
}
\arguments{
\item{pop.pheno}{list of all phenotype information}

\item{index.wt}{economic weights of selection index method, its length should equals to the number of traits}
}
\value{
the weight of index selection
}
\description{
Build date: Feb 8, 2020
Last update: Feb 8, 2020
}
\examples{
pop <- getpop(nind = 100, from = 1, ratio = 0.1)
pop.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)

effs <-
    cal.effs(pop.geno = pop.geno,
             cal.model = "A",
             num.qtn.tr1 = c(2, 6, 10),
             sd.tr1 = c(0.4, 0.2, 0.02, 0.02, 0.02, 0.02, 0.02, 0.001),
             dist.qtn.tr1 = rep("normal", 6),
             prob.tr1 = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5),
             shape.tr1 = c(1, 1, 1, 1, 1, 1),
             scale.tr1 = c(1, 1, 1, 1, 1, 1),
             multrait = TRUE,
             num.qtn.trn = matrix(c(18, 10, 10, 20), 2, 2),
             sd.trn = diag(c(1, 0.5)),
             qtn.spot = rep(0.1, 10),
             maf = 0, 
             verbose = TRUE)
             
pop.pheno <-
    phenotype(effs = effs,
              FR = NULL, 
              pop = pop,
              pop.geno = pop.geno,
              pos.map = NULL,
              h2.tr1 = c(0.3, 0.1, 0.05, 0.05, 0.05, 0.01),
              gnt.cov = matrix(c(1, 2, 2, 15), 2, 2),
              h2.trn = c(0.3, 0.5), 
              sel.crit = "pheno", 
              pop.total = pop, 
              sel.on = TRUE,
              inner.env = NULL, 
              verbose = TRUE)
pop <- pop.pheno$pop
pop.pheno$pop <- NULL 
b <- cal.idx(pop.pheno = pop.pheno, index.wt = c(0.5, 0.5))
}
\author{
Dong Yin
}
