% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selects.r
\name{cal.multi}
\alias{cal.multi}
\title{Multiple trait selection method}
\usage{
cal.multi(pop, decr, sel.multi, index.wt, index.tdm, goal, pass.perc,
  pop.pheno)
}
\arguments{
\item{pop}{population information of generation, family index, within-family index, index, sire, dam, sex, phenotpye}

\item{decr}{whether to sorting with descreasing}

\item{sel.multi}{selection method of multi-trait with "tdm", "indcul", "index"}

\item{index.wt}{economic weights of selection index method}

\item{index.tdm}{ndex represents which trait is being selected. NOT CONTROL BY USER}

\item{goal}{goal of the trait}

\item{pass.perc}{percentage of expected excellent individuals}

\item{pop.pheno}{list of all phenotype information}
}
\value{
individual indice sorted by scores
}
\description{
Build date: Nov 14, 2018
Last update: Oct 13, 2019
}
\examples{
pop <- getpop(nind = 100, from = 1, ratio = 0.1)
pop.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)

effs <-
    cal.effs(pop.geno = pop.geno,
             cal.model = "A",
             num.qtn.tr1 = c(2, 6, 10),
             sd.tr1 = c(0.4, 0.2, 0.02, 0.02, 0.02, 0.02, 0.02, 0.001),
             dist.qtn.tr1 = rep("normal", 6),
             eff.unit.tr1 = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5),
             shape.tr1 = c(1, 1, 1, 1, 1, 1),
             scale.tr1 = c(1, 1, 1, 1, 1, 1),
             multrait = TRUE,
             num.qtn.trn = matrix(c(18, 10, 10, 20), 2, 2),
             sd.trn = diag(c(1, 0.5)),
             qtn.spot = rep(0.1, 10),
             maf = 0, 
             verbose = TRUE)
             
pop.pheno <-
    phenotype(effs = effs,
              FR = NULL, 
              pop = pop,
              pop.geno = pop.geno,
              pos.map = NULL,
              h2.tr1 = c(0.3, 0.1, 0.05, 0.05, 0.05, 0.01),
              gnt.cov = matrix(c(1, 2, 2, 15), 2, 2),
              h2.trn = c(0.3, 0.5), 
              sel.crit = "pheno", 
              pop.total = pop, 
              sel.on = TRUE,
              inner.env = NULL, 
              verbose = TRUE)
pop <- pop.pheno$pop
pop.pheno$pop <- NULL 

f1 <- grep(pattern = "TBV|TGV|pheno|ebv", x = names(pop), value = TRUE)
pheno <- subset(pop, select = f1)                  
goal.perc <- 0.9
goal <- (1 + goal.perc) * apply(pheno, 2, mean)

ind.ordered <- cal.multi(pop = pop, decr = TRUE, sel.multi = "index",
    index.wt = c(0.5, 0.5), index.tdm = 1, goal = goal, 
    pass.perc = 0.9, pop.pheno = pop.pheno)
str(ind.ordered)
}
\author{
Dong Yin
}
