% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.r
\name{cal.pheno}
\alias{cal.pheno}
\title{Calculate for phenotype}
\usage{
cal.pheno(fr = NULL, info.eff = NULL, h2 = NULL, num.ind = NULL,
  var.pheno = NULL, verbose = TRUE)
}
\arguments{
\item{fr}{list of fixed effects and random effects}

\item{info.eff}{list of phenotype decomposition}

\item{h2}{heritability vector of the trait, every elements are corresponding to a, d, aXa, aXd, dXa, dXd respectively}

\item{num.ind}{population size}

\item{var.pheno}{phenotype variace}

\item{verbose}{whether to print detail}
}
\value{
list of phenotype
}
\description{
Build date: Nov 4, 2019
Last update: Nov 4, 2019
}
\examples{
pop <- getpop(nind = 100, from = 1, ratio = 0.1)
pop.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)
geno <- geno.cvt1(pop.geno)
num.ind <- ncol(geno)
h2 <- 0.3
effs <-
    cal.effs(pop.geno = pop.geno,
             cal.model = "A",
             num.qtn.tr1 = c(2, 6, 10),
             sd.tr1 = c(0.4, 0.2, 0.02, 0.02, 0.02, 0.02, 0.02, 0.001),
             dist.qtn.tr1 = rep("normal", 6),
             prob.tr1 = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5),
             shape.tr1 = c(1, 1, 1, 1, 1, 1),
             scale.tr1 = c(1, 1, 1, 1, 1, 1),
             multrait = FALSE,
             num.qtn.trn = matrix(c(18, 10, 10, 20), 2, 2),
             sd.trn = diag(c(1, 0.5)),
             qtn.spot = rep(0.1, 10),
             maf = 0, 
             verbose = TRUE)
        
info.eff <- cal.gnt(geno = geno, h2 = h2, effs = effs, 
    sel.on = TRUE, inner.env = NULL, verbose = TRUE)

var.pheno <- var(info.eff$ind.a) / h2
pheno.list <- cal.pheno(fr = NULL, info.eff = info.eff, h2 = h2,
    num.ind = num.ind, var.pheno = var.pheno, verbose = TRUE)
str(pheno.list)
}
\author{
Dong Yin
}
