% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selects.r
\name{cal.sing}
\alias{cal.sing}
\title{Single trait selection method}
\usage{
cal.sing(pop, decr, sel.sing, cor.r)
}
\arguments{
\item{pop}{population information of generation, family index, within-family index, index, sire, dam, sex, phenotype}

\item{decr}{whether to sorting with decreasing}

\item{sel.sing}{selection method of single trait with "ind", "fam", "infam", "comb"}

\item{cor.r}{average relationship of population}
}
\value{
individual indice sorted by scores
}
\description{
Build date: Nov 14, 2018
Last update: Jul 31, 2019
}
\examples{
pop <- getpop(nind = 100, from = 1, ratio = 0.1)
pop.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)

effs <-
    cal.effs(pop.geno = pop.geno,
             cal.model = "A",
             num.qtn.tr1 = c(2, 6, 10),
             sd.tr1 = c(0.4, 0.2, 0.02, 0.02, 0.02, 0.02, 0.02, 0.001),
             dist.qtn.tr1 = rep("normal", 6),
             prob.tr1 = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5),
             shape.tr1 = c(1, 1, 1, 1, 1, 1),
             scale.tr1 = c(1, 1, 1, 1, 1, 1),
             multrait = FALSE,
             num.qtn.trn = matrix(c(18, 10, 10, 20), 2, 2),
             sd.trn = diag(c(1, 0.5)),
             qtn.spot = rep(0.1, 10),
             maf = 0, 
             verbose = TRUE)

pop.pheno <-
    phenotype(effs = effs,
              FR = NULL, 
              pop = pop,
              pop.geno = pop.geno,
              pos.map = NULL,
              h2.tr1 = c(0.3, 0.1, 0.05, 0.05, 0.05, 0.01),
              gnt.cov = matrix(c(1, 2, 2, 15), 2, 2),
              h2.trn = c(0.3, 0.5), 
              sel.crit = "pheno", 
              pop.total = pop, 
              sel.on = TRUE, 
              inner.env = NULL, 
              verbose = TRUE)
pop <- pop.pheno$pop
pop.pheno$pop <- NULL

cor.r <- cal.r(pop, pop)
ind.ordered <- cal.sing(pop = pop, decr = TRUE, sel.sing = "comb", cor.r = cor.r)
str(ind.ordered)
}
\author{
Dong Yin
}
