% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{call.HIBLUP}
\alias{call.HIBLUP}
\title{Call HIBLUP function}
\usage{
call.HIBLUP(
  pheno,
  cols = NULL,
  covar = NULL,
  fixed_eff = NULL,
  random_eff = NULL,
  geno = NULL,
  geno.id = NULL,
  map = NULL,
  pedigree = NULL,
  mode = "A"
)
}
\arguments{
\item{pheno}{the full phenotype data}

\item{cols}{the column which users need}

\item{covar}{the covariate list}

\item{fixed_eff}{the fixed effect list}

\item{random_eff}{the randm effect list}

\item{geno}{the genotype data}

\item{geno.id}{the id of genotyped individual}

\item{map}{the marker information map}

\item{pedigree}{the pedigree data}

\item{mode}{'A' or 'AD' for additive effect model or additive and dominance model}
}
\value{
gebv
}
\description{
Build date: Jul 16, 2021
Last update: Jul 16, 2021
}
\examples{
# with data(hidata) in HIBLUP software
# gebv <- call.HIBLUP(pheno = pheno, pedigree = pedigree)
}
\author{
Dong Yin
}
