% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.r
\name{check.map}
\alias{check.map}
\title{Get map containing block and recombination information}
\usage{
check.map(input.map = NULL, num.marker = NULL, len.block = 5e+07)
}
\arguments{
\item{input.map}{map that should be input, the marker number should be consistent in both map file and genotype data}

\item{num.marker}{number of markers of genotype matrix}

\item{len.block}{length of every blocks}
}
\value{
a map containing block and recombination information
}
\description{
Build date: Nov 14, 2018
Last update: Jul 30, 2019
}
\examples{
data(simdata)
nmrk <- nrow(input.map)
pos.map <- check.map(input.map = input.map, num.marker = nmrk, len.block = 5e7)
str(pos.map)
}
\author{
Dong Yin
}
