% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Breeding.R
\name{complan}
\alias{complan}
\title{Read and compare breeding plans}
\usage{
complan(SP = NULL, selPath = NULL, verbose = TRUE)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{selPath}{the path of breeding_plan}

\item{verbose}{whether to print details}
}
\value{
None
}
\description{
Build date: Nov 14, 2018
Last update: Jul 31, 2019
}
\examples{
pop.env <- list(
  F1 = list( # fixed effect 1
    level = c("1", "2"),
    eff = list(tr1 = c(50, 30), tr2 = c(50, 30))
  ), 
  F2 = list( # fixed effect 2
    level = c("d1", "d2", "d3"),
    eff = list(tr1 = c(10, 20, 30), tr2 = c(10, 20, 30))
  ),
  R1 = list( # random effect 1
    level = c("l1", "l2", "l3"),
    ratio = list(tr1 = 0.1, tr2 = 0.1)
  )
)
SP <- param.simer(
  out.format = "plink",
  outpath = tempdir(),
  qtn.num = diag(10, 2),
  pop.env = pop.env,
  phe.model = list(
    tr1 = "T1 = A + F1 + F2 + R1 + E",
    tr2 = "T2 = A + F1 + F2 + R1 + E"
  )
)

SP <- simer(SP)

selPath <- system.file("extdata", "04breeding_plan", package = "simer")

# follow code can be run only if you have installd hiblup package
}
\author{
Dong Yin
}
