% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_activity_dist.R
\name{est_activity_dist}
\alias{est_activity_dist}
\title{Estimate the activity distribution of a network}
\usage{
est_activity_dist(branch_fun, N)
}
\arguments{
\item{branch_fun}{the brancing function (output from est_branching_fun())}

\item{N}{the number of nodes in the network}
}
\value{
the expected change in activity level at the next time-step
}
\description{
Generate the probability distribution over active states
based on the branching function by solving the master equation
}
