% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_branching_fun.R
\name{est_branching_fun}
\alias{est_branching_fun}
\title{Generate the branching function}
\usage{
est_branching_fun(W_E, W_I, k, alpha, s_vec = seq(5e-04, 1, 5e-04))
}
\arguments{
\item{W_E}{the effective inhibitory weight}

\item{W_I}{the effective excitatory weight}

\item{k}{mean degree of network (N*p)}

\item{alpha}{the proportion of inhibitory neurons}

\item{s_vec}{the activity levels (proportion of active nodes) to sample
for the branching function}
}
\value{
the expected change in activity level at the next time-step
}
\description{
Estimate the branching function for the given network params
}
