% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_criticality.R
\name{est_criticality}
\alias{est_criticality}
\title{Estimate the third parameter which will lead to near-criticality}
\usage{
est_criticality(alpha = NULL, e_eff = NULL, i_eff = NULL)
}
\arguments{
\item{alpha}{The proportion of nodes that are inhibitory}

\item{e_eff}{The effective excitatory weight}

\item{i_eff}{The effective inhibitory weight}
}
\description{
Estimate the third parameter which will lead to near-criticality
}
