% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Genotype.r
\name{generate.map}
\alias{generate.map}
\title{Generate Marker information}
\usage{
generate.map(pop.marker = NULL, num.chr = 18, len.chr = 1.5e+08)
}
\arguments{
\item{pop.marker}{the number of markers}

\item{num.chr}{the number of chromosomes}

\item{len.chr}{the length of chromosomes}
}
\value{
marker information
}
\description{
Build date: Nov 9, 2020
Last update: Nov 9, 2020
}
\examples{
pop.map <- generate.map(pop.marker = 1e4)
str(pop.map)
}
\author{
Dong Yin
}
