% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Phenotype.r
\name{generate.pop}
\alias{generate.pop}
\title{Generate population according to number of individuals}
\usage{
generate.pop(pop.ind = 100, from = 1, ratio = 0.5, gen = 1)
}
\arguments{
\item{pop.ind}{number of the individuals in a population}

\item{from}{initial index of the population}

\item{ratio}{ratio of males in a population}

\item{gen}{generation ID of the population}
}
\value{
population information
}
\description{
Build date: Nov 14, 2018
Last update: Mar 23, 2022
}
\examples{
pop <- generate.pop(pop.ind = 100)
head(pop)
}
\author{
Dong Yin
}
