% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.r
\name{geno.cvt}
\alias{geno.cvt}
\title{Convert genotype matrix from (0, 1) to (0, 1, 2)}
\usage{
geno.cvt(pop.geno)
}
\arguments{
\item{pop.geno}{genotype matrix of (0, 1)}
}
\value{
genotype matrix of (0, 1, 2)
}
\description{
Build date: Nov 14, 2018
Last update: Jul 30, 2019
}
\examples{
num.marker <- 48353
num.ind <- 100
geno1 <- genotype(num.marker = num.marker, num.ind = num.ind, verbose = TRUE)
geno2 <- geno.cvt(pop.geno = geno1)
geno1[1:5, 1:10]
geno2[1:5, 1:5]
}
\author{
Dong Yin
}
