% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{geno.cvt2}
\alias{geno.cvt2}
\title{Convert genotype matrix from (0, 1, 2) to (0, 1)}
\usage{
geno.cvt2(pop.geno)
}
\arguments{
\item{pop.geno}{genotype matrix of (0, 1, 2)}
}
\value{
genotype matrix of (0, 1)
}
\description{
Build date: Jul 11, 2020
Last update: Jul 11, 2020
}
\examples{
data(simdata)
geno1 <- geno.cvt1(rawgeno)
geno1[1:5, 1:5]
geno <- geno.cvt2(pop.geno = geno1)
geno[1:5, 1:10]
}
\author{
Dong Yin
}
