% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Genotype.r
\name{genotype}
\alias{genotype}
\title{Generate genotype and editing genotype}
\usage{
genotype(SP = NULL, ncpus = 0, verbose = TRUE)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{ncpus}{the number of threads}

\item{verbose}{whether to print detail}
}
\value{
a genotype matrix with block and map information
}
\description{
Build date: Nov 14, 2018
Last update: Feb 20, 2022
}
\examples{
SP <- param.geno(pop.marker = 1e4, pop.ind = 1e2)
SP <- genotype(SP)
SP$geno$pop.geno$gen1[1:5, 1:5]
}
\author{
Dong Yin
}
