% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.r
\name{getgmt}
\alias{getgmt}
\title{Get true position in the genotype matrix}
\usage{
getgmt(index, incols = 2)
}
\arguments{
\item{index}{the position of individual in the genotype matrix}

\item{incols}{the column number of an individual in the input genotype matrix, it can be 1 or 2}
}
\value{
sub-genotype matrix
}
\description{
Build date: June 11, 2020
Last update: June 11, 2020
}
\examples{
index <- c(1:2, 5:6)
gmt <- getgmt(index = index, incols = 2)
}
