% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Reproduces.r
\name{getindex}
\alias{getindex}
\title{Get output index of phenotype and pedigree}
\usage{
getindex(count.ind, out.gen)
}
\arguments{
\item{count.ind}{a vector with population size of every generations}

\item{out.gen}{a vector with generations needing outputting}
}
\value{
individuals indice need outputting
}
\description{
Build date: Nov 14, 2018
Last update: Aug 1, 2019
}
\examples{
count.ind <- c(100, 200, 400, 800)
out.gen <- 2:4
idx <- getindex(count.ind = count.ind, out.gen = out.gen)
str(idx)
}
\author{
Dong Yin
}
