% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Reproduces.r
\name{getsd}
\alias{getsd}
\title{Get selected sires and dams}
\usage{
getsd(ind.ordered, pop, num.sir, num.dam)
}
\arguments{
\item{ind.ordered}{individual ID after sorting}

\item{pop}{population information}

\item{num.sir}{number of sires}

\item{num.dam}{number of dams}
}
\value{
sire ID and dam ID after selection
}
\description{
Build date: Nov 14, 2018
Last update: Apr 30, 2020
}
\examples{
basepop <- getpop(nind = 100, from = 1, ratio = 0.5)
ind.ordered <- sample.int(100)

ind.stay <- getsd(ind.ordered = ind.ordered, pop = basepop, 
                  num.sir = 10, num.dam = 40)
str(ind.stay)
}
\author{
Dong Yin
}
