% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Genotype.r
\name{GxG.network}
\alias{GxG.network}
\title{Generate genetic interaction effect combination}
\usage{
GxG.network(pop.map = NULL, qtn.pos = 1:10, qtn.model = "A:D")
}
\arguments{
\item{pop.map}{the marker information data.}

\item{qtn.pos}{the index of QTNs in the map.}

\item{qtn.model}{the genetic effect model of QTN.}
}
\value{
a dataframe of genetic interaction effect
}
\description{
Build date: Mar 19, 2022
Last update: Mar 19, 2022
}
\examples{
pop.map <- generate.map(pop.marker = 1e4)
GxG.net <- GxG.network(pop.map)
head(GxG.net)
}
\author{
Dong Yin
}
