% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_network.R
\name{init_network}
\alias{init_network}
\title{Initialize a random, homogenous network}
\usage{
init_network(n, p, alpha = 0.1, e_eff, i_eff, normal_weigths = F)
}
\arguments{
\item{n}{The number of nodes in the network}

\item{p}{The probability that node i feeds into node j (and visa-versa)}

\item{alpha}{The proportion of neurons that are inhibitory}

\item{e_eff}{The effective excitatory weight}

\item{i_eff}{The effective inhibitory weight}

\item{k}{(optional) The expected degree of the network}
}
\description{
Initialize a network
}
