% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_network_input.R
\name{init_network_ss}
\alias{init_network_ss}
\alias{init_network_ss}
\title{Simulate the dynamics of a network instance without input in order to get
a reasonable steady state}
\usage{
init_network_ss(net_ss, in_symbol, t_steps = 1)

init_network_ss(net_ss, in_symbol, t_steps = 1)
}
\arguments{
\item{net_ss}{a network and activity vector that represent steady state
(the output from init_network())}

\item{in_symbol}{a vector of length N of values in [-1,1] that represents
added input to the specified node. A value of 0 means there is no extra input}

\item{net}{a network (the output from init_network()) to simulate}

\item{n_steps}{the number of time steps to simulate}

\item{s_1}{the proporation of active nodes at time step 1}

\item{n_steps}{the number of time steps to simulate}
}
\description{
Simulate the dynamics of a network instance without input in order to get
a reasonable steady state

Simulate the dynamics of a steady state network instance with input
}
