% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Genotype.r
\name{input.geno}
\alias{input.geno}
\title{Input genotype column by column if markers are dense}
\usage{
input.geno(bigmtr, mtr, ed, mrk.dense)
}
\arguments{
\item{bigmtr}{total genotype matrix}

\item{mtr}{genotype matrix should be inputting}

\item{ed}{index of the last column in each process}

\item{mrk.dense}{whether markers are dense, it is TRUE when sequencing data}
}
\value{
none
}
\description{
Build date: Nov 14, 2018
Last update: Jul 30, 2019
}
\examples{
bigmtr <- bigmemory::big.matrix(nrow = 1e4, ncol = 2e2, type = 'char')
bigmtr[1:5, 1:5]
options(bigmemory.typecast.warning=FALSE)
mtr <- matrix(0, 1e4, 1e2)
input.geno(bigmtr = bigmtr, mtr = mtr, ed = ncol(mtr), mrk.dense = FALSE)
bigmtr[1:5, 1:5]
}
\author{
Dong Yin
}
