% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{input.map}
\alias{input.map}
\title{Map file from outside in simdata}
\format{list}
\usage{
data(simdata)
}
\description{
Map file from outside in simdata
}
\examples{
data(simdata)
dim(input.map)
head(input.map)
}
\keyword{datasets}
