% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.Json.R
\name{JsonModel}
\alias{JsonModel}
\title{The main function of Constructing Model and Index}
\usage{
JsonModel(
  jsonFile,
  buildModel = TRUE,
  buildIndex = TRUE,
  ncpus = 10,
  verbose = TRUE
)
}
\arguments{
\item{jsonFile}{the path of json file}

\item{buildModel}{whether to build EBV model}

\item{buildIndex}{whether to build Selection Index}

\item{ncpus}{the number of threads}

\item{verbose}{whether to print detail.}
}
\value{
report list
}
\description{
Build date: Aug 16, 2021
Last update: Aug 16, 2021
}
\examples{
jsonFile <- system.file("extdata", "demo2.json", package = "simer")
# aa <- JsonModel(jsonFile = jsonFile)
}
\author{
Dong Yin
}
