% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{load_if_installed}
\alias{load_if_installed}
\title{Installation checking}
\usage{
load_if_installed(package)
}
\arguments{
\item{package}{the package name.}
}
\value{
none.
}
\description{
Check if the software is installed.
}
\details{
Build date: Oct 22, 2018
Last update: Apr 30, 2022
}
\author{
Dong Yin, Lilin Yin, Haohao Zhang, and Xiaolei Liu
}
\keyword{internal}
