% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Logging.R
\name{logging.initialize}
\alias{logging.initialize}
\title{Logging initialization}
\usage{
logging.initialize(module, outpath)
}
\arguments{
\item{module}{the module name.}

\item{outpath}{the path of output files, Simer writes files only if outpath is not 'NULL'.}
}
\value{
none.
}
\description{
Initialize the logging process.
}
\details{
Build date: Jul 11, 2020
Last update: Apr 28, 2022
}
\author{
Dong Yin
}
