% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Logging.R
\name{logging.log}
\alias{logging.log}
\title{Print or write log}
\usage{
logging.log(
  ...,
  file = NULL,
  sep = " ",
  fill = FALSE,
  labels = NULL,
  verbose = TRUE
)
}
\arguments{
\item{...}{R objects}

\item{file}{A connection, or a character string naming the file to print to. If "" (the default), cat prints to the standard output connection, the console unless redirected by sink. If it is "|cmd", the output is piped to the command given by ‘cmd’, by opening a pipe connection.}

\item{sep}{a character vector of strings to append after each element.}

\item{fill}{a logical or (positive) numeric controlling how the output is broken into successive lines.}

\item{labels}{character vector of labels for the lines printed. Ignored if fill is FALSE.}

\item{verbose}{whether to print detail}
}
\value{
print information
}
\description{
Print or write log
}
\examples{
logging.log('simer')
}
