% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Logging.R
\name{logging.print}
\alias{logging.print}
\title{Print things into file}
\usage{
logging.print(x, file = NULL, append = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{a matrix or a list}

\item{file}{output file name}

\item{append}{ogical. If TRUE, output will be appended to file; otherwise, it will overwrite the contents of file}

\item{verbose}{whether to print details}
}
\value{
print in the screen
}
\description{
Build date: Feb 7, 2020
Last update: Feb 7, 2020
by using base::print
}
\examples{
x <- list(a = "a", b = "b")
simer.print(x)
}
\author{
Dong Yin
}
