% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{make_line}
\alias{make_line}
\title{make line}
\usage{
make_line(string, width, linechar = " ", align = "center", margin = 1)
}
\description{
Build date: Dec 12, 2018
Last update: Dec 12, 2018
}
\author{
Haohao Zhang
}
\keyword{internal}
