% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Reproduces.r
\name{mate.3waycro}
\alias{mate.3waycro}
\title{three-way cross process}
\usage{
mate.3waycro(SP, ncpus = 0, verbose = TRUE)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{ncpus}{the number of threads}

\item{verbose}{whether to print detail}
}
\value{
population information and genotype matrix of population after three-way cross process
}
\description{
Build date: Apr 11, 2022
Last update: Apr 11, 2022
}
\examples{
SP <- param.annot(qtn.num = 10)
SP <- param.geno(SP = SP, pop.marker = 1e4, pop.ind = 1e2)
SP <- param.pheno(SP = SP, pop.ind = 100)
SP <- param.sel(SP = SP, sel.single = "comb")
SP <- param.reprod(SP = SP, prog = 2)
SP <- annotation(SP)
SP <- genotype(SP)
SP <- phenotype(SP)
# three different breeds are cut by sex
SP$pheno$pop$gen1$sex <- rep(c(1, 2, 1), c(30, 30, 40))
SP <- selects(SP)
SP <- mate.3waycro(SP)
str(SP)
}
\author{
Dong Yin
}
