% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Reproduces.r
\name{mate.clone}
\alias{mate.clone}
\title{Clone process}
\usage{
mate.clone(SP, ncpus = 0, verbose = TRUE)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{ncpus}{the number of threads}

\item{verbose}{whether to print detail}
}
\value{
population information and genotype matrix of population after clone process
}
\description{
Build date: Nov 14, 2018
Last update: Apr 10, 2022
}
\examples{
SP <- param.annot(qtn.num = 10)
SP <- param.geno(SP = SP, pop.marker = 1e4, pop.ind = 1e2)
SP <- param.pheno(SP = SP, pop.ind = 100)
SP <- param.sel(SP = SP, sel.single = "comb")
SP <- param.reprod(SP = SP, prog = 2)
SP <- annotation(SP)
SP <- genotype(SP)
SP <- phenotype(SP)
SP <- selects(SP)
SP <- mate.clone(SP)
str(SP)
}
\author{
Dong Yin
}
