% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduces.r
\name{mate.selfpol}
\alias{mate.selfpol}
\title{Self-pollination}
\usage{
mate.selfpol(pop1, pop1.geno, ind.stay, num.prog, ratio)
}
\arguments{
\item{pop1}{population information of population1}

\item{pop1.geno}{genotype matrix of population1}

\item{ind.stay}{selected individuals regarded as parents}

\item{num.prog}{litter size of dams}

\item{ratio}{ratio of males in all individuals}
}
\value{
population information and genotype matrix of population after self-pollination process
}
\description{
Build date: Nov 14, 2018
Last update: Aug 1, 2019
}
\examples{
basepop <- getpop(nind = 100, from = 1, ratio = 0.1)
basepop.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)
pop.selfpol <- mate.selfpol(pop1 = basepop, pop1.geno = basepop.geno,
                 ind.stay = basepop$index, num.prog = 2, ratio = 0.5)
pop <- pop.selfpol$pop
geno <- pop.selfpol$geno
str(pop)
str(geno)
}
\author{
Dong Yin
}
