% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduces.r
\name{mate.singcro}
\alias{mate.singcro}
\title{Single cross process}
\usage{
mate.singcro(pop1, pop2, pop1.geno.id, pop2.geno.id, pop1.geno, pop2.geno,
  incols = 2, ind.stay, num.prog, ratio)
}
\arguments{
\item{pop1}{population information of population1}

\item{pop2}{population information of population2}

\item{pop1.geno.id}{genotype id of population1}

\item{pop2.geno.id}{genotype id of population2}

\item{pop1.geno}{genotype matrix of population1}

\item{pop2.geno}{genotype matrix of population2}

\item{incols}{the column number of an individual in the input genotype matrix, it can be 1 or 2}

\item{ind.stay}{selected individuals regarded as parents}

\item{num.prog}{litter size of dams}

\item{ratio}{ratio of males in all individuals}
}
\value{
population information and genotype matrix of population after single cross process
}
\description{
Build date: Nov 14, 2018
Last update: Aug 1, 2019
}
\examples{
pop1 <- getpop(nind = 100, from = 1, ratio = 0.1)
pop2 <- getpop(nind = 100, from = 101, ratio = 0.1)
pop1.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)
pop2.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)
ind.stay <- list(sir = pop1$index, dam = pop2$index)
pop.singcro <- mate.singcro(pop1 = pop2, pop2 = pop2,
    pop1.geno.id = pop1$index, pop2.geno.id = pop2$index, 
    pop1.geno = pop1.geno, pop2.geno = pop2.geno,
    ind.stay = ind.stay, num.prog = 2, ratio = 0.5)
pop <- pop.singcro$pop
geno <- pop.singcro$geno
str(pop)
str(geno)
}
\author{
Dong Yin
}
