% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Reproduces.r
\name{mate.userped}
\alias{mate.userped}
\title{User-spicified cross process}
\usage{
mate.userped(SP, ncpus = 0, verbose = TRUE)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{ncpus}{the number of threads}

\item{verbose}{whether to print detail}
}
\value{
population information and genotype matrix of population after User-spicified cross process
}
\description{
Build date: Apr 12, 2022
Last update: Apr 12, 2022
}
\examples{
SP <- param.annot(qtn.num = 10)
SP <- param.geno(SP = SP, pop.marker = 1e4, pop.ind = 1e2)
SP <- param.pheno(SP = SP, pop.ind = 100)
SP <- param.reprod(SP = SP, reprod.way = "userped")
SP <- annotation(SP)
SP <- genotype(SP)
SP <- phenotype(SP)
SP <- mate.userped(SP)
str(SP)
}
\author{
Dong Yin
}
