% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{mkl_env}
\alias{mkl_env}
\title{MKL environment}
\usage{
mkl_env(exprs, threads = 1)
}
\arguments{
\item{exprs}{the expression.}

\item{threads}{the number of threads used, if NULL, (logical core number - 1) is automatically used.}
}
\value{
none.
}
\description{
Run code in the MKL environment.
}
\details{
Build date: Oct 22, 2018
Last update: Apr 30, 2022
}
\author{
Dong Yin, Lilin Yin, Haohao Zhang, and Xiaolei Liu
}
\keyword{internal}
