% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Breeding.R
\name{num2num}
\alias{num2num}
\title{Convert string number to numeric number}
\usage{
num2num(nums)
}
\arguments{
\item{nums}{string number}
}
\value{
numeric number
}
\description{
Build date: Nov 14, 2018
Last update: Jul 31, 2019
}
\examples{
nums <- "1,2,4:7"
num <- num2num(nums)
num
}
\author{
Dong Yin
}
