% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Parameter.r
\name{param.annot}
\alias{param.annot}
\title{Generate parameters for annotation data simulation}
\usage{
param.annot(SP = NULL, ...)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{...}{one or more parameter(s) for map simulation}
}
\value{
a list of map simulation parameters
}
\description{
Build date: Feb 24, 2022
Last update: Mar 23, 2022
}
\examples{
SP <- param.annot(qtn.num = 10)
str(SP)
}
\author{
Dong Yin
}
