% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Parameter.r
\name{param.global}
\alias{param.global}
\title{Generate parameters for global options}
\usage{
param.global(SP = NULL, ...)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{...}{one or more parameter(s) for global options}
}
\value{
a list of global options
}
\description{
Build date: Apr 16, 2022
Last update: Apr 16, 2022
}
\examples{
SP <- param.global(out = "simer")
str(SP)
}
\author{
Dong Yin
}
