% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Parameter.r
\name{param.pheno}
\alias{param.pheno}
\title{Generate parameters for phenotype data simulation}
\usage{
param.pheno(SP = NULL, ...)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{...}{one or more parameter(s) for phenotype simulation}
}
\value{
a list of phenotype simulation parameters
}
\description{
Build date: Feb 21, 2022
Last update: Mar 23, 2022
}
\examples{
SP <- param.pheno(phe.model = list(tr1 = "T1 = A + E"))
str(SP)
}
\author{
Dong Yin
}
