% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Parameter.r
\name{param.reprod}
\alias{param.reprod}
\title{Generate parameters for reproduction}
\usage{
param.reprod(SP = NULL, ...)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{...}{one or more parameter(s) for reproduction}
}
\value{
a list of selection parameters
}
\description{
Build date: Apr 6, 2022
Last update: Apr 6, 2022
}
\examples{
SP <- param.reprod(reprod.way = "randmate")
str(SP)
}
\author{
Dong Yin
}
