% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Parameter.r
\name{param.sel}
\alias{param.sel}
\title{Generate parameters for selection}
\usage{
param.sel(SP = NULL, ...)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{...}{one or more parameter(s) for selection}
}
\value{
a list of selection parameters
}
\description{
Build date: Apr 6, 2022
Last update: Apr 6, 2022
}
\examples{
SP <- param.sel(sel.single = "comb")
str(SP)
}
\author{
Dong Yin
}
