% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{pop.geno}
\alias{pop.geno}
\title{Raw genotype matrix from outside in simdata}
\format{
matrix
}
\usage{
data(simdata)
}
\description{
Raw genotype matrix from outside in simdata
}
\examples{
data(simdata)
dim(pop.geno)
head(pop.geno)
}
\keyword{datasets}
