% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{pop.map}
\alias{pop.map}
\title{Map file from outside in simdata}
\format{
list
}
\usage{
data(simdata)
}
\description{
Map file from outside in simdata
}
\examples{
data(simdata)
dim(pop.map)
head(pop.map)
}
\keyword{datasets}
