% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{print_accomplished}
\alias{print_accomplished}
\title{Accomplishment}
\usage{
print_accomplished(width = 60, verbose = TRUE)
}
\arguments{
\item{width}{the width of the message.}

\item{verbose}{whether to print detail.}
}
\value{
none.
}
\description{
Print accomplishment information.
}
\details{
Build date: Aug 30, 2017
Last update: Apr 30, 2022
}
\author{
Dong Yin, Lilin Yin, Haohao Zhang, and Xiaolei Liu
}
\keyword{internal}
