% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{print_info}
\alias{print_info}
\title{Simer information}
\usage{
print_info(
  welcome = NULL,
  title = NULL,
  short_title = NULL,
  logo = NULL,
  version = NULL,
  authors = NULL,
  contact = NULL,
  linechar = "=",
  width = NULL,
  verbose = TRUE
)
}
\arguments{
\item{welcome}{welcome text, for example: "Welcom to <Packagename>".}

\item{title}{long text to introduct package.}

\item{short_title}{short label, top-left of logo.}

\item{logo}{logo.}

\item{version}{short label, bottom-right of logo.}

\item{authors}{authors of software.}

\item{contact}{email or website.}

\item{width}{banner width.}

\item{verbose}{whether to print detail.}

\item{line}{1, 2, or char.}
}
\description{
Print R Package information, include title, short_title, logo, version, authors, contact.
}
\details{
Build date: Oct 22, 2018
Last update: Apr 30, 2022
}
\examples{
welcome <- "Welcome to SIMER"
title   <- "Data Simulation for Life Science and Breeding"
authors <- c("Designed and Maintained by Dong Yin, Xuanning Zhang,
              Lilin Yin, Haohao Zhang, and Xiaolei Liu", 
             "Contributors: Zhenshuang Tang, Jingya Xu, Xinyun Li, 
              Mengjin Zhu, Xiaohui Yuan, Shuhong Zhao")
contact <- "Contact: xiaoleiliu@mail.hzau.edu.cn"
logo_s  <- c(" ____ ___ __  __ _____ ____  ", 
             "/ ___|_ _|  \\\\/  | ____|  _ \\\\ ", 
             "\\\\___ \\\\| || |\\\\/| |  _| | |_) |", 
             " ___) | || |  | | |___|  _ < ", 
             "|____/___|_|  |_|_____|_| \\\\_\\\\")
print_info(welcome = welcome, title = title, logo = logo_s, authors = authors, 
           contact = contact, linechar = '=', width = 70)
}
\author{
Dong Yin and Haohao Zhang
}
\keyword{internal}
