% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{rawgeno}
\alias{rawgeno}
\title{Raw genotype matrix from outside in simdata}
\format{big.matrix}
\usage{
data(simdata)
}
\description{
Raw genotype matrix from outside in simdata
}
\examples{
data(simdata)
dim(rawgeno)
head(rawgeno)
}
\keyword{datasets}
