% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selects.r
\name{read.selgeno}
\alias{read.selgeno}
\title{Read sel_ind_geno to get selection condition}
\usage{
read.selgeno(pop = NULL, selPath = NULL, out = NULL)
}
\arguments{
\item{pop}{total population information}

\item{selPath}{the path of select_scheme}

\item{out}{path of output files}
}
\value{
None
}
\description{
Build date: Nov 14, 2018
Last update: Jul 31, 2019
}
\examples{
basepop <- getpop(nind = 100, from = 1, ratio = 0.1)
basepop.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)
effs <-
    cal.effs(pop.geno = basepop.geno,
             cal.model = "A",
             num.qtn.tr1 = c(2, 6, 10),
             var.tr1 = c(0.4, 0.2, 0.02, 0.02, 0.02, 0.02, 0.02, 0.001),
             dist.qtn.tr1 = rep("normal", 6),
             eff.unit.tr1 = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5),
             shape.tr1 = c(1, 1, 1, 1, 1, 1),
             scale.tr1 = c(1, 1, 1, 1, 1, 1),
             multrait = FALSE, # single trait
             num.qtn.trn = matrix(c(18, 10, 10, 20), 2, 2),
             eff.sd = diag(c(1, 0.5)),
             qtn.spot = rep(0.1, 10),
             maf = 0, 
             verbose = TRUE)
pop.pheno <-
    phenotype(effs = effs,
              pop = basepop,
              pop.geno = basepop.geno,
              pos.map = NULL,
              h2.tr1 = c(0.3, 0.1, 0.05, 0.05, 0.05, 0.01),
              gnt.cov = matrix(c(1, 2, 2, 15), 2, 2),
              env.cov = matrix(c(10, 5, 5, 100), 2, 2),
              sel.crit = "pheno", 
              pop.total = basepop, 
              sel.on = TRUE, 
              inner.env = NULL, 
              verbose = TRUE)
str(basepop)
basepop <- set.pheno(basepop, pop.pheno, sel.crit = "pheno")
str(basepop)
selPath <- system.file("extdata", "01select_scheme", package = "simer")
out.index <- read.selgeno(pop = basepop, selPath = selPath, out = NULL)
}
\author{
Dong Yin
}
