% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{remove_bigmatrix}
\alias{remove_bigmatrix}
\title{Big.matrix removing}
\usage{
remove_bigmatrix(x, desc_suffix = ".geno.desc", bin_suffix = ".geno.bin")
}
\arguments{
\item{x}{the filename of big.matrix.}

\item{desc_suffix}{the suffix of description file of big.matrix.}

\item{bin_suffix}{the suffix of binary file of big.matrix.}
}
\value{
TRUE or FALSE
}
\description{
Remove big.matrix safely.
}
\details{
Build date: Aug 8, 2019
Last update: Apr 30, 2022
}
\examples{
library(bigmemory)
mat <- filebacked.big.matrix(
     nrow = 10,
     ncol = 10,
     init = 0,
     type = 'char',
     backingpath = ".",
     backingfile = 'simer.geno.bin',
     descriptorfile = 'simer.geno.desc')

remove_bigmatrix(x = "simer")
}
\author{
Haohao Zhang and Dong Yin
}
