% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{rule_wrap}
\alias{rule_wrap}
\title{wrap text to multiple line, align left, right or center.}
\usage{
rule_wrap(string, width, align = "center", linechar = " ")
}
\description{
Build date: Oct 22, 2018
Last update: Dec 12, 2018
by using base::strwarp.
}
\author{
Haohao Zhang
}
\keyword{internal}
