% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Selects.r
\name{sel.core}
\alias{sel.core}
\title{Get core population ID and genotype}
\usage{
sel.core(
  pop.sel = NULL,
  core.stay = NULL,
  refresh = rep(0.6, 2),
  keep.max.gen = rep(3, 2),
  incols = 2,
  pop.total,
  pop.geno.curr,
  pop.geno.core
)
}
\arguments{
\item{pop.sel}{ID of the sires and the dams passing selection}

\item{core.stay}{ID the core sires and dams}

\item{refresh}{refresh ratio of core population of sires and dams, only used in ps > 1}

\item{keep.max.gen}{if ps <= 1, fraction selected in selection of males and females; if ps > 1, ps is number of selected males and females}

\item{incols}{the column number of an individual in the input genotype matrix, it can be 1 or 2}

\item{pop.total}{total population information}

\item{pop.geno.curr}{genotype matrix of current population}

\item{pop.geno.core}{genotype matrix of core population}
}
\value{
core population list
}
\description{
Build date: May 2, 2020
Last update: May 2, 2020
}
\examples{
# no example for now
}
\author{
Dong Yin
}
