% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Breeding.R
\name{sel.idx}
\alias{sel.idx}
\title{Select individual ID according to scheme}
\usage{
sel.idx(scheme, pop)
}
\arguments{
\item{scheme}{breeding scheme for genotyping or phenotyping}

\item{pop}{population information}
}
\value{
index vector
}
\description{
Build date: Jan 13, 2020
Last update: Jan 13, 2020
}
\examples{
scheme <- data.frame(
    generation = c("1", "2", "3"), 
    family_index = rep("1:3,7",3),
    within_family_index = rep("all", 3), 
    sex = rep("1", 3), 
    stringsAsFactors = FALSE
    )
pop <- getpop(nind = 100, from = 1, ratio = 0.5)
pop$gen <- rep(1:5, each = 20)
out.idx <- sel.idx(scheme = scheme, pop = pop)
}
\author{
Dong Yin
}
