% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Selects.r
\name{selects}
\alias{selects}
\title{Select individuals by combination of selection method and criterion}
\usage{
selects(SP = NULL, verbose = TRUE)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{verbose}{whether to print detail}
}
\value{
individual indice sorted by scores
}
\description{
Build date: Sep 8, 2018
Last update: Apr 4, 2022
}
\examples{
SP <- param.annot(qtn.num = 10)
SP <- param.geno(SP = SP, pop.marker = 1e4, pop.ind = 1e2)
SP <- param.pheno(SP = SP, pop.ind = 100)
SP <- param.sel(SP = SP, sel.single = "comb")
SP <- annotation(SP)
SP <- genotype(SP)
SP <- phenotype(SP)
SP <- selects(SP)
str(SP$sel$pop.sel$gen1)
}
\author{
Dong Yin
}
