% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.R
\name{simer.Data.cHIBLUP}
\alias{simer.Data.cHIBLUP}
\title{simer.Data.cHIBLUP: The function of calling HIBLUP software of C version
Author: Dong Yin
Build date: June 28, 2021
Last update: Apr 21, 2022}
\usage{
simer.Data.cHIBLUP(
  SP = NULL,
  jsonList = NULL,
  mode = "A",
  vc.method = "AI",
  ncpus = 10,
  verbose = TRUE
)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{jsonList}{the list of json parameters}

\item{mode}{'A' or 'AD', Additive effect model or Additive and Dominance mode}

\item{vc.method}{default is 'AI', the method of calucating variance components in HIBLUP}

\item{ncpus}{the number of threads}

\item{verbose}{whether to print detail.}
}
\value{
a list of gebv
}
\description{
simer.Data.cHIBLUP: The function of calling HIBLUP software of C version
Author: Dong Yin
Build date: June 28, 2021
Last update: Apr 21, 2022
}
\examples{
jsonFile <- system.file("extdata", "demo2.json", package = "simer")
jsonList <- rjson::fromJSON(file = jsonFile)
# gebvs <- simer.Data.cHIBLUP(SP = SP)
# gebvs <- simer.Data.cHIBLUP(jsonList = jsonList)
}
