% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.R
\name{simer.Data.Env}
\alias{simer.Data.Env}
\title{To find appropriate fixed effects, covariates, and random effects
Author: Dong Yin
Build date: July 17, 2021
Last update: Apr 21, 2022}
\usage{
simer.Data.Env(
  jsonList = NULL,
  header = TRUE,
  sep = "\\t",
  ncpus = 10,
  verbose = TRUE
)
}
\arguments{
\item{jsonList}{the list of json parameters}

\item{header}{the header of file}

\item{sep}{the separator of file}

\item{ncpus}{the number of threads}

\item{verbose}{whether to print detail.}
}
\value{
the best effects for EBV model
}
\description{
To find appropriate fixed effects, covariates, and random effects
Author: Dong Yin
Build date: July 17, 2021
Last update: Apr 21, 2022
}
\examples{
jsonFile <- system.file("extdata", "demo2.json", package = "simer")
jsonList <- rjson::fromJSON(file = jsonFile)
# jsonList <- simer.Data.Env(jsonList = jsonList)
}
