% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.R
\name{simer.Data.Geno}
\alias{simer.Data.Geno}
\title{simer.Data.Geno: Data quality control of genotype data
Author: Dong Yin
Build date: June 3, 2021
Last update: Aug 22, 2021}
\usage{
simer.Data.Geno(
  fileMVP,
  filePed = NULL,
  out = "simer",
  genoType = "char",
  filterGeno = NULL,
  filterHWE = NULL,
  filterMind = NULL,
  filterMAF = NULL,
  ncpus = 0,
  verbose = TRUE
)
}
\arguments{
\item{fileMVP}{the prefix of MVP file}

\item{filePed}{the name of pedigree file}

\item{out}{the name of output file}

\item{genoType}{type parameter in bigmemory, genotype data. The default is char, it is highly recommended *NOT* to modify this parameter.}

\item{filterGeno}{threshold of sample miss rate}

\item{filterHWE}{threshold of Hardy-Weinberg Test}

\item{filterMind}{threshold of variant miss rate}

\item{filterMAF}{threshold of Minor Allele Frequency}

\item{ncpus}{The number of threads used, if NULL, (logical core number - 1) is automatically used}

\item{verbose}{whether to print detail.}
}
\value{
Output file:
<out>.geno.desc
<out>.geno.bin
<out>.geno.ind
}
\description{
simer.Data.Geno: Data quality control of genotype data
Author: Dong Yin
Build date: June 3, 2021
Last update: Aug 22, 2021
}
\examples{
# simer.Data.Ped needs genotype data
out <- tempfile("outfile")
fileMVP <- system.file("extdata", "01bigmemory", "demo", package = "simer")
filePed <- system.file("extdata", "pedigree.txt", package = "simer")
# simer.Data.Geno(fileMVP=fileMVP, filePed=filePed, genoType='char', out=out, 
  # filterGeno=0.1, filterHWE=0.001, filterMind=0.1, filterMAF=0.05,
  # ncpus=0, verbose=TRUE)
}
