% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.R
\name{simer.Data.Impute}
\alias{simer.Data.Impute}
\title{simer.Data.Impute: Impute the missing value
Author: Dong Yin
Build date: May 26, 2021
Last update: Sep 18, 2021}
\usage{
simer.Data.Impute(
  fileMVP = NULL,
  fileBed = NULL,
  out = NULL,
  maxLine = 10000,
  ncpus = 0,
  verbose = TRUE
)
}
\arguments{
\item{fileMVP}{Genotype in MVP format}

\item{fileBed}{Genotype in PLINK binary format}

\item{out}{the name of output file}

\item{maxLine}{number of SNPs, only used for saving memory when calculate kinship matrix}

\item{ncpus}{The number of threads used, if NULL, (logical core number - 1) is automatically used}

\item{verbose}{whether to print detail.}
}
\value{
genotype data after imputing
}
\description{
simer.Data.Impute: Impute the missing value
Author: Dong Yin
Build date: May 26, 2021
Last update: Sep 18, 2021
}
\examples{
fileMVP <- system.file("extdata", "01bigmemory", "demo", package = "simer")
fileMVPimp <- simer.Data.Impute(fileMVP = fileMVP)
}
