% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.R
\name{simer.Data.Map}
\alias{simer.Data.Map}
\title{simer.Data.Map: To check map file}
\usage{
simer.Data.Map(
  map,
  out = "simer",
  cols = 1:5,
  header = TRUE,
  sep = "\\t",
  verbose = TRUE
)
}
\arguments{
\item{map}{the name of map file or map object(data.frame or matrix).}

\item{out}{the name of output file.}

\item{cols}{selected columns.}

\item{header}{whether the file contains header.}

\item{sep}{seperator of the file.}

\item{verbose}{whether to print detail.}
}
\value{
Output file:
<out>.map
}
\description{
checking map file.
}
\details{
Build date: Sep 12, 2018
Last update: July 25, 2022
}
\examples{
# Get map path
mapPath <- system.file("extdata", "01bigmemory", "demo.geno.map", package = "simer")

# Check map data
simer.Data.Map(mapPath, tempfile("outfile"))
}
\author{
Haohao Zhang and Dong Yin
}
