% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.R
\name{simer.Data.MVP2MVP}
\alias{simer.Data.MVP2MVP}
\title{simer.Data.MVP2MVP: Change the path of MVP files
Author: Dong Yin
Build date: May 26, 2021
Last update: May 26, 2021}
\usage{
simer.Data.MVP2MVP(fileMVP, genoType = "char", out = "simer", verbose = TRUE)
}
\arguments{
\item{fileMVP}{the prefix of MVP file}

\item{genoType}{type parameter in bigmemory, genotype data. The default is char, it is highly recommended *NOT* to modify this parameter.}

\item{out}{the name of output file}

\item{verbose}{whether to print detail.}
}
\value{
Output file:
<out>.geno.desc
<out>.geno.bin
<out>.geno.ind
<out>.geno.map
}
\description{
simer.Data.MVP2MVP: Change the path of MVP files
Author: Dong Yin
Build date: May 26, 2021
Last update: May 26, 2021
}
\examples{
# simer.Data.Ped needs genotype data
out <- tempfile("outfile")
mvpPath <- system.file("extdata", "01bigmemory", "demo", package = "simer")

# simer.Data.MVP2MVP(mvpPath)
}
