% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.R
\name{simer.Data.Ped}
\alias{simer.Data.Ped}
\title{simer.Data.Ped: To check pedigree file
Author: LiLin Yin and Dong Yin
Build date: May 6, 2021
Last update: June 3, 2021}
\usage{
simer.Data.Ped(
  filePed,
  fileMVP = NULL,
  out = NULL,
  standardID = FALSE,
  fileSir = NULL,
  fileDam = NULL,
  exclThres = 0.01,
  assignThres = 0.005,
  header = TRUE,
  sep = "\\t",
  ncpus = 0,
  verbose = TRUE
)
}
\arguments{
\item{filePed}{the name of pedigree file need correcting}

\item{fileMVP}{Genotype in MVP format}

\item{out}{the name of output file}

\item{standardID}{whether kid id is 15-character standard}

\item{fileSir}{the file name of candidate sires}

\item{fileDam}{the file name of candidate dams}

\item{exclThres}{if conflict ratio is more than exclThres, exclude this parent}

\item{assignThres}{if conflict ratio is less than assignThres, assign this parent to the individual}

\item{header}{whether the file contains header}

\item{sep}{separator of the file}

\item{ncpus}{The number of threads used, if NULL, (logical core number - 1) is automatically used}

\item{verbose}{whether to print detail.}
}
\value{
Output file:
<out>.report.txt
<out>.error.txt
<out>.qc.txt
}
\description{
simer.Data.Ped: To check pedigree file
Author: LiLin Yin and Dong Yin
Build date: May 6, 2021
Last update: June 3, 2021
}
\examples{
# simer.Data.Ped needs genotype data
out <- tempfile("outfile")
fileMVP <- system.file("extdata", "01bigmemory", "demo", package = "simer")
simer.Data.MVP2MVP(fileMVP, out = out)

filePed <- system.file("extdata", "pedigree.txt", package = "simer")
simer.Data.Ped(filePed = filePed, out = out, standardID = FALSE)
}
