% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.R
\name{simer.Data.Pheno}
\alias{simer.Data.Pheno}
\title{simer.Data.Pheno: Data quality control of phenotype data
Author: Haohao Zhang and Dong Yin
Build date: June 13, 2021
Last update: July 8, 2021}
\usage{
simer.Data.Pheno(
  filePhe,
  filePed = NULL,
  out = NULL,
  planPhe = NULL,
  pheCols = NULL,
  header = TRUE,
  sep = "\\t",
  missing = c(NA, "NA", "Na", ".", "-", "NAN", "nan", "na", "N/A", "n/a", "<NA>", "",
    "-9", 9999),
  verbose = TRUE
)
}
\arguments{
\item{filePhe}{the phenotype files, it can be a vector}

\item{filePed}{the pedigree files, it can be a vector}

\item{out}{the name prefix of output file}

\item{planPhe}{breeding plans about phenotype}

\item{pheCols}{the column needing extracting}

\item{header}{the header of file}

\item{sep}{the separator of file}

\item{missing}{the missing value}

\item{verbose}{whether to print detail.}
}
\value{
Output file:
<out>.qc.txt
}
\description{
simer.Data.Pheno: Data quality control of phenotype data
Author: Haohao Zhang and Dong Yin
Build date: June 13, 2021
Last update: July 8, 2021
}
\examples{
out <- tempfile("outfile")
filePhe <- system.file("extdata", "phenotype.txt", package = "simer")
simer.Data.Pheno(filePhe = filePhe, out = out, verbose = TRUE)
}
