% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.R
\name{simer.Data}
\alias{simer.Data}
\title{Simer.Data: To prepare data for Simer package}
\usage{
simer.Data(
  fileMVP = NULL,
  fileBed = NULL,
  filePlinkPed = NULL,
  genoType = "char",
  filterGeno = NULL,
  filterHWE = NULL,
  filterMind = NULL,
  filterMAF = NULL,
  filePed = NULL,
  standardID = FALSE,
  fileSir = NULL,
  fileDam = NULL,
  exclThres = 0.01,
  assignThres = 0.005,
  pedSep = "\\t",
  filePhe = NULL,
  planPhe = NULL,
  pheCols = NULL,
  pheSep = "\\t",
  missing = c(NA, "NA", "-9", 9999),
  SNP.impute = "Major",
  outpath = getwd(),
  out = "simer",
  maxLine = 10000,
  priority = "speed",
  ncpus = 0,
  verbose = TRUE
)
}
\arguments{
\item{fileMVP}{Genotype in MVP format}

\item{fileBed}{Genotype in PLINK binary format}

\item{filePlinkPed}{Genotype in PLINK ped format}

\item{genoType}{type parameter in bigmemory, genotype data. The default is char, it is highly recommended *NOT* to modify this parameter.}

\item{filterGeno}{threshold of sample miss rate}

\item{filterHWE}{threshold of Hardy-Weinberg Test}

\item{filterMind}{threshold of variant miss rate}

\item{filterMAF}{threshold of Minor Allele Frequency}

\item{filePed}{Pedigree, 3-columns or 15 columns pedigree}

\item{standardID}{whether kid id is 15-character standard}

\item{fileSir}{the file name of candidate sires}

\item{fileDam}{the file name of candidate dams}

\item{exclThres}{if conflict ratio is more than exclThres, exclude this parent}

\item{assignThres}{if conflict ratio is less than assignThres, assign this parent to the individual}

\item{pedSep}{the separator of pedigree file}

\item{filePhe}{Phenotype, the first column is taxa name, the subsequent columns are traits}

\item{planPhe}{breeding plans about phenotype}

\item{pheCols}{the column neeeding extracting}

\item{pheSep}{seperator for phenotype file.}

\item{missing}{the missing value}

\item{SNP.impute}{"Left", "Middle", "Right", or NULL for skip impute.}

\item{outpath}{the output path}

\item{out}{prefix of output file name}

\item{maxLine}{number of SNPs, only used for saving memory when calculate kinship matrix}

\item{priority}{"speed" or "memory"}

\item{ncpus}{The number of threads used, if NULL, (logical core number - 1) is automatically used}

\item{verbose}{whether to print detail.}
}
\description{
Build date: May 26, 2021
Last update: Aug 23, 2021
}
\examples{
out <- tempfile("outfile")
filePhe <- system.file("extdata", "phenotype.txt", package = "simer")
filePhe <- simer.Data.Pheno(filePhe = filePhe, out = out, verbose = TRUE)
}
\author{
Dong Yin
}
