% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.R
\name{simer.Data.SELIND}
\alias{simer.Data.SELIND}
\title{simer.SELIND: The function of General Selection Index
Author: Dong Yin
Build date: Aug 26, 2021
Last update: Apr 21, 2022}
\usage{
simer.Data.SELIND(jsonList = NULL, ncpus = 10, verbose = TRUE)
}
\arguments{
\item{jsonList}{the list of json parameters}

\item{ncpus}{the number of threads}

\item{verbose}{whether to print detail.}
}
\value{
phenotype economic weight
}
\description{
simer.SELIND: The function of General Selection Index
Author: Dong Yin
Build date: Aug 26, 2021
Last update: Apr 21, 2022
}
\examples{
jsonFile <- system.file("extdata", "demo2.json", package = "simer")
jsonList <- rjson::fromJSON(file = jsonFile)
# jsonList <- simer.Data.SELIND(jsonList = jsonList)
}
\references{
Y. S. Chen, Z. L. Sheng (1988) The Theory of General Selection Index. Genetic Report, 15(3): P185-P190
}
