% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.r
\name{simer}
\alias{simer}
\title{Main function of simer}
\usage{
simer(SP)
}
\arguments{
\item{SP}{a list of all simulation parameters}
}
\value{
a list with simulated population
}
\description{
Build date: Jan 7, 2019
Last update: Apr 17, 2022
}
\examples{
SP <- param.simer(out = "simer")
SP <- simer(SP)
}
\author{
Dong Yin, Lilin Yin, Haohao Zhang and Xiaolei Liu
}
