% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{simer.sample}
\alias{simer.sample}
\title{Sample process with all needed elements}
\usage{
simer.sample(x, size, replace = TRUE, prob = NULL)
}
\arguments{
\item{x}{either a vector of one or more elements from which to choose, or a positive integer}

\item{size}{a non-negative integer giving the number of items to choose}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements of the vector being sampled}
}
\value{
sampled elements with all needed elements
}
\description{
Build date: Jul 11, 2020
Last update: Jul 11, 2020
}
\examples{
sam <- simer.sample(x = paste0("a", 1:3), 100)
sam
unique(sam)
}
\author{
Dong Yin
}
