% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{simer.show.file}
\alias{simer.show.file}
\title{Show file content}
\usage{
simer.show.file(filename = NULL, verbose = TRUE)
}
\arguments{
\item{filename}{filename of a file}

\item{verbose}{whether to print details}
}
\value{
none
}
\description{
Build date: Feb 11, 2020
Last update: Feb 11, 2020
}
\examples{
selPath <- system.file("extdata", "04breeding_plan", package = "simer")
filename <- file.path(selPath, "breeding_plan01.txt")
simer.show.file(filename = filename, verbose = TRUE)
}
\author{
Dong Yin
}
