% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{simer.Version}
\alias{simer.Version}
\title{Print simer Banner}
\usage{
simer.Version(width = 60, verbose = TRUE)
}
\arguments{
\item{width}{the width of the message}

\item{verbose}{whether to print detail.}
}
\value{
version number.
}
\description{
Build date: Aug 30, 2017
Last update: Oct 21, 2019
}
\examples{
simer.Version()
}
\author{
Dong Yin, Lilin Yin, Haohao Zhang, and Xiaolei Liu
}
