% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.r
\name{times}
\alias{times}
\title{Convert number to time character}
\usage{
times(x)
}
\arguments{
\item{x}{time of numeric}
}
\value{
time of character in the format "h", "m", "s"
}
\description{
Build date: Jan 7, 2019
Last update: Aug 1, 2019
}
\examples{
time.range <- 3700
times(time.range)
}
\author{
Dong Yin
}
