% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{userped}
\alias{userped}
\title{User-specific pedigree in simdata}
\format{
list
}
\usage{
data(simdata)
}
\description{
User-specific pedigree in simdata
}
\examples{
data(simdata)
dim(userped)
head(userped)
}
\keyword{datasets}
