% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.r
\name{write.file}
\alias{write.file}
\title{Write files of simer}
\usage{
write.file(pop, geno, map, out.geno.index, out.pheno.index, seed.map, out,
  out.format, verbose)
}
\arguments{
\item{pop}{population information of generation, family index, within-family index, index, sire, dam, sex, phenotpye}

\item{geno}{genotype matrix of population}

\item{map}{map information of markers}

\item{out.geno.index}{indice of individuals outputting genotype}

\item{out.pheno.index}{indice of individuals outputting phenotype}

\item{seed.map}{random seed of map file}

\item{out}{path of output files}

\item{out.format}{format of output, "numeric" or "plink"}

\item{verbose}{whether to print detail}
}
\value{
None
}
\description{
Build date: Jan 7, 2019
Last update: Aug 1, 2019
}
\examples{
\donttest{
data(simdata)
nmrk <- nrow(input.map)
pos.map <- check.map(input.map = input.map, num.marker = nmrk, len.block = 5e7)
basepop <- getpop(nind = 40, from = 1, ratio = 0.1)
basepop.geno <- rawgeno
basepop.geno <- as.big.matrix(basepop.geno)
effs <-
    cal.effs(pop.geno = basepop.geno,
             cal.model = "A",
             num.qtn.tr1 = c(2, 6, 10),
             var.tr1 = c(0.4, 0.2, 0.02, 0.02, 0.02, 0.02, 0.02, 0.001),
             dist.qtn.tr1 = rep("normal", 6),
             eff.unit.tr1 = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5),
             shape.tr1 = c(1, 1, 1, 1, 1, 1),
             scale.tr1 = c(1, 1, 1, 1, 1, 1),
             multrait = FALSE,
             num.qtn.trn = matrix(c(18, 10, 10, 20), 2, 2),
             eff.sd = diag(c(1, 0.5)),
             qtn.spot = rep(0.1, 10),
             maf = 0, 
             verbose = TRUE)
str(basepop)
pop.pheno <-
    phenotype(effs = effs,
              pop = basepop,
              pop.geno = basepop.geno,
              pos.map = NULL,
              h2.tr1 = c(0.3, 0.1, 0.05, 0.05, 0.05, 0.01),
              gnt.cov = matrix(c(1, 2, 2, 15), 2, 2),
              env.cov = matrix(c(10, 5, 5, 100), 2, 2),
              sel.crit = "pheno", 
              pop.total = basepop, 
              sel.on = TRUE, 
              inner.env = NULL, 
              verbose = TRUE)
basepop <- set.pheno(pop = basepop, pop.pheno = pop.pheno, sel.crit = "pheno")
idx <- basepop$index
seed.map <- 888888
# convert (0, 1) to (0, 1, 2)
basepop.geno <- geno.cvt(basepop.geno)
basepop.geno <- as.big.matrix(basepop.geno)
write.file(pop = basepop, geno = basepop.geno, map = pos.map, 
    out.geno.index = idx, out.pheno.index = idx, seed.map = seed.map, 
    out = tempdir(), out.format = "numeric", verbose = TRUE)
file.remove(file.path(tempdir(), "geno_id.txt"))
file.remove(file.path(tempdir(), "map.txt"))
file.remove(file.path(tempdir(), "pedigree.txt"))
file.remove(file.path(tempdir(), "phenotype.txt"))
}
}
\author{
Dong Yin
}
