% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.r
\name{build.cov}
\alias{build.cov}
\title{To bulid correlation of variables}
\usage{
build.cov(df = NULL, mu = rep(0, nrow(Sigma)), Sigma, tol = 1e-06)
}
\arguments{
\item{df}{data.frame without correlation}

\item{mu}{means of the variables}

\item{Sigma}{covariance matrix of variables}

\item{tol}{tolerance (relative to largest variance) for numerical 
lack of positive-definiteness in Sigma.}
}
\value{
data.frame with correlaion
}
\description{
Build date: Oct 10, 2019
Last update: Oct 10, 2019
}
\examples{
Sigma <- matrix(c(14, 10, 10, 15), 2, 2)
Sigma
df <- cbind(rnorm(100), 0)
df <- as.data.frame(df)
names(df) <- paste0("tr", 1:ncol(df))
df.cov <- build.cov(df, Sigma = Sigma)
var(df.cov)
}
\references{
B. D. Ripley (1987) Stochastic Simulation. Wiley. Page 98
}
\author{
Dong Yin and R
}
