% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.r
\name{build.cov}
\alias{build.cov}
\title{To bulid correlation of variables}
\usage{
build.cov(mat, mu = rep(0, nrow(Sigma)), Sigma, tol = 1e-06)
}
\arguments{
\item{mat}{matrix without correlation}

\item{mu}{means of the variables}

\item{Sigma}{covariance matrix of variables}

\item{tol}{tolerance (relative to largest variance) for numerical 
lack of positive-definiteness in Sigma.}
}
\value{
matrix with correlaion
}
\description{
Build date: Oct 10, 2019
Last update: Oct 10, 2019
}
\examples{
Sigma <- matrix(c(14, 10, 10, 15), 2, 2)
Sigma
mat <- cbind(rnorm(100), rnorm(100))
mat.cov <- build.cov(mat, Sigma = Sigma)
var(mat.cov)
}
\references{
B. D. Ripley (1987) Stochastic Simulation. Wiley. Page 98
}
\author{
Dong Yin and R
}
