% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.r
\name{cal.eff}
\alias{cal.eff}
\title{Calculate for genetic effects vector of selected markers}
\usage{
cal.eff(num.qtn, eff.sd, dist.qtn, eff.unit, shape, scale)
}
\arguments{
\item{num.qtn}{number of QTN}

\item{eff.sd}{standard deviation of different effects}

\item{dist.qtn}{distribution of QTN's effects with options: "normal", "geometry" and "gamma"}

\item{eff.unit}{unit effect of geometric distribution}

\item{shape}{shape of gamma distribution}

\item{scale}{scale of gamma distribution}
}
\value{
genetic effects vector of selected markers
}
\description{
Build date: Nov 14, 2018
Last update: Jul 30, 2019
}
\examples{
num.qtn <- c(2, 6, 10) # three qtn groups
eff.sd <- c(0.4, 0.2, 0.02) # three variance of qtn group
eff <- cal.eff(num.qtn = num.qtn, eff.sd = eff.sd, dist.qtn = "normal",
        eff.unit = 0.5, shape = 1, scale = 1)
str(eff)

num.qtn <- sum(num.qtn)
eff.sd <- sum(eff.sd)
eff <- cal.eff(num.qtn = num.qtn, eff.sd = eff.sd, dist.qtn = "normal",
        eff.unit = 0.5, shape = 1, scale = 1)
str(eff)
}
\author{
Dong Yin
}
