% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Genotype.r
\name{cal.eff}
\alias{cal.eff}
\title{Calculate for genetic effects vector of selected markers}
\usage{
cal.eff(
  qtn.num = 10,
  qtn.dist = "normal",
  qtn.sd = 1,
  qtn.prob = 0.5,
  qtn.shape = 1,
  qtn.scale = 1,
  qtn.shape1 = 1,
  qtn.shape2 = 1,
  qtn.ncp = 0
)
}
\arguments{
\item{qtn.num}{number of QTN}

\item{qtn.dist}{distribution of QTN's effects with options: "normal", "geometry", "gamma", and "beta"}

\item{qtn.sd}{standard deviation of different effects}

\item{qtn.prob}{unit effect of geometric distribution}

\item{qtn.shape}{shape of gamma distribution}

\item{qtn.scale}{scale of gamma distribution}

\item{qtn.shape1}{non-negative parameters of the Beta distribution}

\item{qtn.shape2}{non-negative parameters of the Beta distribution}

\item{qtn.ncp}{non-centrality parameter}
}
\value{
genetic effects vector of selected markers
}
\description{
Build date: Nov 14, 2018
Last update: Mar 16, 2022
}
\examples{
eff <- cal.eff(qtn.num = 10, qtn.dist = "norm")
str(eff)
}
\author{
Dong Yin
}
