% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.r
\name{cal.effs}
\alias{cal.effs}
\title{Calculate for genetic effects list of selected markers}
\usage{
cal.effs(pop.geno = NULL, cal.model = "A", num.qtn.tr1 = c(2, 6, 10),
  sd.tr1 = c(0.4, 0.2, 0.02, 0.02, 0.02, 0.02, 0.02, 0.001),
  dist.qtn.tr1 = rep("normal", 6), eff.unit.tr1 = c(0.5, 0.5, 0.5, 0.5,
  0.5, 0.5), shape.tr1 = c(1, 1, 1, 1, 1, 1), scale.tr1 = c(1, 1, 1, 1,
  1, 1), multrait = FALSE, num.qtn.trn = matrix(c(18, 10, 10, 20), 2,
  2), sd.trn = matrix(c(1, 0, 0, 0.5), 2, 2), qtn.spot = rep(0.1, 10),
  maf = 0, verbose = TRUE)
}
\arguments{
\item{pop.geno}{genotype of population, a individual has two columns}

\item{cal.model}{phenotype models with the options: "A", "AD", "ADI"}

\item{num.qtn.tr1}{integer or integer vector, the number of QTN in the trait1}

\item{sd.tr1}{standard deviation of different effects, the last 5 vector elements are corresponding to d, aXa, aXd, dXa, dXd respectively and the rest elements are corresponding to a}

\item{dist.qtn.tr1}{distributions of the QTN effects with the options: "normal", "geometry" and "gamma", vector elements are corresponding to a, d, aXa, aXd, dXa, dXd respectively}

\item{eff.unit.tr1}{unit effect of geometric distribution of the trait1, vector elements are corresponding to a, d, aXa, aXd, dXa, dXd respectively}

\item{shape.tr1}{shape of gamma distribution of the trait1, vector elements are corresponding to a, d, aXa, aXd, dXa, dXd respectively}

\item{scale.tr1}{scale of gamma distribution of the trait1, vector elements are corresponding to a, d, aXa, aXd, dXa, dXd respectively}

\item{multrait}{whether to apply multiple traits, TRUE represents applying, FALSE represents not}

\item{num.qtn.trn}{QTN distribution matrix, diagonal elements are total QTN number of the trait, non-diagonal elements are QTN number of overlap QTN between two traits}

\item{sd.trn}{a matrix with the standard deviation of the QTN effects}

\item{qtn.spot}{QTN probability in every block}

\item{maf}{Minor Allele Frequency, marker selection range is from  maf to 0.5}

\item{verbose}{whether to print detail}
}
\value{
a list with number of overlap markers, selected markers, effect of markers
}
\description{
Build date: Sep 11, 2018
Last update: Jul 30, 2019
}
\examples{
basepop.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)
effs <-
    cal.effs(pop.geno = basepop.geno,
             cal.model = "A",
             num.qtn.tr1 = c(200, 200, 100),
             sd.tr1 = c(0.07, 0.07, 0.07, 0.07, 0.07, 0.07, 0.07, 0.03),
             dist.qtn.tr1 = rep("normal", 6),
             eff.unit.tr1 = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5),
             shape.tr1 = c(1, 1, 1, 1, 1, 1),
             scale.tr1 = c(1, 1, 1, 1, 1, 1),
             multrait = FALSE,
             num.qtn.trn = matrix(c(400, 100, 100, 400), 2, 2),
             sd.trn = matrix(c(0.07, 0, 0, 0.07), 2, 2),
             qtn.spot = rep(0.1, 10),
             maf = 0, 
             verbose = TRUE)
str(effs)
}
\author{
Dong Yin
}
