% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.r
\name{cal.FR}
\alias{cal.FR}
\title{Calculate for fixed effects and random effects}
\usage{
cal.FR(pop = NULL, FR, var.pheno = NULL, pop.env = NULL,
  verbose = TRUE)
}
\arguments{
\item{pop}{population information}

\item{FR}{list of fixed effects, random effects, and their combination}

\item{var.pheno}{phenotype variances of all traits}

\item{pop.env}{R environment of population information}

\item{verbose}{whether to print detail}
}
\value{
list of fixed effects and random effects
}
\description{
Build date: Nov 1, 2019
Last update: Nov 1, 2019
}
\examples{
pop <- getpop(100, 1, 0.5)
a <- sample(c("a1", "a2", "a3"), 100, replace = TRUE)
b <- sample(c("b1", "b2", "b3"), 100, replace = TRUE)
pop$a <- a # load your fixed  effects
pop$b <- b # load your random effects
pop.env <- environment()

# combination of fixed effects
cmb.fix <- list(tr1 = c("mu", "gen", "sex"), # trait 1
                tr2 = c("mu", "diet", "season")) # trait 2
            
fix <- list( 
         mu = list(level = "mu", eff = 2),  				      
        gen = FALSE,         
        sex = list(level = c("1", "2"), eff = c(0.5, 0.3)), 
       diet = list(level = c("d1", "d2", "d3"), eff = c(0.1, 0.2, 0.3)),
     season = list(level = c("s1", "s2", "s3", "s4"), eff = c(0.1, 0.2, 0.3, 0.2)), 
          a = list(level = c("a1", "a2", "a3"), eff = c(0.1, 0.2, 0.3))) 

# combination and ralation of random effects
tr1 <- list(rn = c("sir", "PE"), ratio = c(0.01, 0.03), 
            cr = matrix(c(1, 0.5, 0.5, 1), 2, 2))
tr2 <- list(rn = c("dam", "litter"), ratio = c(0.01, 0.03), 
            cr = matrix(c(1, 0.5, 0.5, 1), 2, 2))          
cmb.rand <- list(tr1 = tr1, tr2 = tr2)   
      
rand <- list(
        sir = list(mean = 0, sd = 1),		      
        dam = list(mean = 0, sd = 1),	       
         PE = list(level = c("p1", "p2", "p3"), eff = c(0.01, 0.02, 0.03)), 
     litter = list(level = c("l1", "l2"), eff = c(0.01, 0.02)),    
          b = list(level = c("b1", "b2", "b3"), eff = c(0.01, 0.02, 0.03)))
  
FR <- list(cmb.fix = cmb.fix, fix = fix, cmb.rand = cmb.rand, rand = rand)
fr <- cal.FR(pop = pop, FR = FR, var.pheno = c(10, 10), pop.env = pop.env, verbose = TRUE)
str(fr) 	          	         	          	         
}
\author{
Dong Yin
}
