% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.R
\name{checkEnv}
\alias{checkEnv}
\title{Environmental factor checking}
\usage{
checkEnv(data, envName)
}
\arguments{
\item{data}{data needing check.}

\item{envName}{the environmental factor name within the data.}
}
\value{
data without environmental factors of wrong level
}
\description{
Check the levels of environmental factors
}
\details{
Build date: Sep 10, 2021
Last update: Apr 28, 2022
}
\examples{
data <- data.frame(a = rep(1, 3), b = 1:3, c = c(1, 1, 5))
envName <- c("a", "b", "c")

# It contains warnings
# data <- checkEnv(data = data, envName = envName)
}
\author{
Dong Yin
}
