% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Breeding.R
\name{complan}
\alias{complan}
\title{Read and compare breeding plans}
\usage{
complan(
  SP = NULL,
  simls = NULL,
  FR = NULL,
  index.wt = c(0.5, 0.5),
  decr = TRUE,
  selPath = NULL,
  verbose = TRUE
)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{simls}{a list with all simer result information}

\item{FR}{list of fixed effects, random effects, and their combination}

\item{index.wt}{economic weights of selection index method, its length should equals to the number of traits}

\item{decr}{whether to sorting with decreasing}

\item{selPath}{the path of breeding_plan}

\item{verbose}{whether to print details}
}
\value{
None
}
\description{
Build date: Nov 14, 2018
Last update: Jul 31, 2019
}
\examples{
pop.env <- list(
  F1 = list(
    level = c("1", "2"),
    eff = list(tr1 = c(50, 30), tr2 = c(50, 30))
  ), 
  F2 = list(
    level = c("d1", "d2", "d3"),
    eff = list(tr1 = c(10, 20, 30), tr2 = c(10, 20, 30))
  ),
  R1 = list(
    level = c("l1", "l2", "l3"),
    ratio = list(tr1 = 0.1, tr2 = 0.1)
  )
)
SP <- param.simer(
  qtn.num = diag(10, 2),
  pop.env = pop.env,
  phe.model = list(
    tr1 = "T1 = A + F1 + F2 + R1 + E",
    tr2 = "T2 = A + F1 + F2 + R1 + E"
  )
)

SP <- simer(SP, outpath = "/home/yindong/Desktop")

selPath <- system.file("extdata", "04breeding_plan", package = "simer")

# follow code can be run only if you have installd hiblup package
# goal.plan <- complan(simls = simer.list, FR = FR, index.wt = c(0.5, 0.5), 
#                      decr = TRUE, selPath = selPath, verbose = TRUE)
# str(goal.plan)
}
\author{
Dong Yin
}
