% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Genotype.r
\name{generate.map}
\alias{generate.map}
\title{Marker information}
\usage{
generate.map(
  species = NULL,
  pop.marker = NULL,
  num.chr = 18,
  len.chr = 1.5e+08
)
}
\arguments{
\item{species}{the species of genetic map.}

\item{pop.marker}{the number of markers.}

\item{num.chr}{the number of chromosomes.}

\item{len.chr}{the length of chromosomes.}
}
\value{
a data frame with marker information.
}
\description{
Generate map data with marker information.
}
\details{
Build date: Mar 19, 2022
Last update: Apr 28, 2022
}
\examples{
pop.map <- generate.map(pop.marker = 1e4)
str(pop.map)
}
\author{
Dong Yin
}
