% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Genotype.r
\name{geno.cvt1}
\alias{geno.cvt1}
\title{Genotype code convertor 1}
\usage{
geno.cvt1(pop.geno)
}
\arguments{
\item{pop.geno}{genotype matrix of (0, 1).}
}
\value{
genotype matrix of (0, 1, 2).
}
\description{
Convert genotype matrix from (0, 1) to (0, 1, 2).
}
\details{
Build date: Nov 14, 2018
Last update: Apr 28, 2022
}
\examples{
\donttest{
SP <- param.geno(pop.marker = 1e4, pop.ind = 1e2, incols = 2)
SP <- genotype(SP)
geno1 <- SP$geno$pop.geno$gen1
geno2 <- geno.cvt1(geno1)
geno1[1:6, 1:4]
geno2[1:6, 1:2]
}
}
\author{
Dong Yin
}
