% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Genotype.r
\name{geno.cvt2}
\alias{geno.cvt2}
\title{Convert genotype matrix from (0, 1, 2) to (0, 1)}
\usage{
geno.cvt2(pop.geno)
}
\arguments{
\item{pop.geno}{genotype matrix of (0, 1, 2)}
}
\value{
genotype matrix of (0, 1)
}
\description{
Build date: Jul 11, 2020
Last update: Jul 11, 2020
}
\examples{
SP <- param.geno(pop.marker = 1e4, pop.ind = 1e2, incols = 1)
SP <- genotype(SP)
geno1 <- SP$geno$pop.geno$gen1
geno2 <- geno.cvt2(geno1)
geno1[1:6, 1:2]
geno2[1:6, 1:4]
}
\author{
Dong Yin
}
