% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.r
\name{genotype}
\alias{genotype}
\title{Generate genotype and editing genotype}
\usage{
genotype(rawgeno = NULL, geno = NULL, num.marker = NULL,
  num.ind = NULL, prob = c(0.5, 0.5), blk.rg = NULL,
  recom.spot = NULL, range.hot = 4:6, range.cold = 1:5,
  rate.mut = 1e-08, verbose = TRUE)
}
\arguments{
\item{rawgeno}{genotype matrix from outside}

\item{geno}{genotype matrix need dealing with}

\item{num.marker}{number of markers}

\item{num.ind}{number of individuals}

\item{prob}{weight of "0" and "1" in genotype matrix, the sum of element in vector equals 1}

\item{blk.rg}{represents the started and ended position blocks}

\item{recom.spot}{whether to consider recombination in every blocks}

\item{range.hot}{range of exchages in hot spot block}

\item{range.cold}{range of exchages in cold spot block}

\item{rate.mut}{mutation rate between 1e-8 and 1e-6}

\item{verbose}{whether to print detail}
}
\value{
a genotype matrix with block and map information
}
\description{
Build date: Nov 14, 2018
Last update: Oct 13, 2019
}
\examples{
# get map file and create genotype matrix
data(simdata)
num.marker = nrow(input.map)
num.ind = 100
basepop.geno <- genotype(num.marker = num.marker, num.ind = num.ind, verbose = TRUE)
basepop.geno[1:5, 1:5]

# get block information and recombination information
nmrk <- nrow(basepop.geno)
nind <- ncol(basepop.geno) / 2
pos.map <- check.map(input.map = input.map, num.marker = nmrk, len.block = 5e7)
blk.rg <- cal.blk(pos.map)
recom.spot <- as.numeric(pos.map[blk.rg[, 1], 7])

# genotype matrix after Exchange and Mutation
basepop.geno.em <-
genotype(geno = basepop.geno,
         blk.rg = blk.rg,
         recom.spot = recom.spot,
         range.hot = 4:6,
         range.cold = 1:5,
         rate.mut = 1e-8, 
         verbose = TRUE)
basepop.geno.em[1:5, 1:5]
}
\author{
Dong Yin
}
