% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduces.r
\name{getpop}
\alias{getpop}
\title{Generate population according to number of individauls}
\usage{
getpop(nind, from, ratio, gen = 1)
}
\arguments{
\item{nind}{number of the individuals in a population}

\item{from}{initial index of the population}

\item{ratio}{ratio of males in a population}

\item{gen}{generation ID of the population}
}
\value{
population information
}
\description{
Build date: Nov 14, 2018
Last update: Aug 1, 2019
}
\examples{
basepop <- getpop(nind = 100, from = 1, ratio = 0.1)
str(basepop)
}
\author{
Dong Yin
}
