% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Genotype.r
\name{GxG.network}
\alias{GxG.network}
\title{Genetic interaction network}
\usage{
GxG.network(pop.map = NULL, qtn.pos = 1:10, qtn.model = "A:D")
}
\arguments{
\item{pop.map}{the map data with annotation information.}

\item{qtn.pos}{the index of QTNs in the map data.}

\item{qtn.model}{the genetic model of QTN such as 'A:D'.}
}
\value{
a data frame of genetic interaction effect.
}
\description{
Generate genetic interaction effect combination network.
}
\details{
Build date: Mar 19, 2022
Last update: Apr 28, 2022
}
\examples{
pop.map <- generate.map(pop.marker = 1e4)
GxG.net <- GxG.network(pop.map)
head(GxG.net)
}
\author{
Dong Yin
}
