% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Reproduces.r
\name{IndPerGen}
\alias{IndPerGen}
\title{Calculate the individual number per generation}
\usage{
IndPerGen(
  pop,
  pop.gen = 2,
  ps = c(0.8, 0.8),
  reprod.way = "randmate",
  sex.rate = 0.5,
  prog = 2
)
}
\arguments{
\item{pop}{the base population}

\item{pop.gen}{the generation number of population}

\item{ps}{if ps <= 1, fraction selected in selection of males and females; if ps > 1, ps is number of selected males and females}

\item{reprod.way}{the reproduction way}

\item{sex.rate}{the number of progeny per dam}

\item{prog}{the number of progeny per dam}
}
\value{
the vector containing the individual number per generation
}
\description{
Build date: Apr 12, 2022
Last update: Apr 12, 2022
}
\examples{
pop <- generate.pop(pop.ind = 100)
}
\author{
Dong Yin
}
