% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.Json.R
\name{JsonQC}
\alias{JsonQC}
\title{The main function of Data Quality Control}
\usage{
JsonQC(jsonFile, ncpus = 0, verbose = TRUE)
}
\arguments{
\item{jsonFile}{the path of json file}

\item{ncpus}{the number of threads}

\item{verbose}{whether to print detail.}
}
\value{
report list
}
\description{
Build date: Oct 19, 2020
Last update: May 26, 2021
}
\examples{
jsonFile <- system.file("extdata", "demo1.json", package = "simer")
# JsonQC() needs set work directory at data directory 
# aa <- JsonQC(jsonFile = jsonFile)
}
\author{
Dong Yin
}
