% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Logging.R
\name{logging.print}
\alias{logging.print}
\title{Logging printer}
\usage{
logging.print(x, file = NULL, append = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{a matrix or a list.}

\item{file}{the filename of output file.}

\item{append}{logical. If TRUE, output will be appended to file; otherwise, it will overwrite the contents of file.}

\item{verbose}{whether to print details.}
}
\value{
none.
}
\description{
Print R object information into file.
}
\details{
Build date: Feb 7, 2020
Last update: Apr 28, 2022
}
\examples{
x <- list(a = "a", b = "b")
logging.print(x)
}
\author{
Dong Yin
}
