% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduces.r
\name{mate.randmate}
\alias{mate.randmate}
\title{Random mating process}
\usage{
mate.randmate(pop1, pop1.geno, incols = 2, ind.stay, num.prog, ratio)
}
\arguments{
\item{pop1}{population information of population1}

\item{pop1.geno}{genotype matrix of population1}

\item{incols}{the column number of an individual in the input genotype matrix, it can be 1 or 2}

\item{ind.stay}{selected individuals regarded as parents}

\item{num.prog}{litter size of dams}

\item{ratio}{ratio of males in all individuals}
}
\value{
population information and genotype matrix of population after random mating process
}
\description{
Build date: Nov 14, 2018
Last update: Aug 1, 2019
}
\examples{
basepop <- getpop(nind = 100, from = 1, ratio = 0.1)
basepop.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)
pop.randmate <- mate.randmate(pop1 = basepop, pop1.geno = basepop.geno,
    ind.stay = basepop$index, num.prog = 2, ratio = 0.5)
pop <- pop.randmate$pop
geno <- pop.randmate$geno
str(pop)
str(geno)
}
\author{
Dong Yin
}
