% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduces.r
\name{mate}
\alias{mate}
\title{Mating according to indice of sires and dams}
\usage{
mate(pop.geno, index.sir, index.dam)
}
\arguments{
\item{pop.geno}{genotype matrix of population}

\item{index.sir}{indice of sires}

\item{index.dam}{indice of dams}
}
\value{
genotype matrix after mating
}
\description{
Build date: Nov 14, 2018
Last update: Aug 1, 2019
}
\examples{
pop.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)
index.sir <- rep(c(1, 3, 4, 5, 7, 9), each = 2)
index.dam <- rep(c(11, 12, 13, 14, 16, 18), each = 2)
geno.curr <- mate(pop.geno = pop.geno, index.sir = index.sir,
                 index.dam = index.dam)
str(geno.curr)
}
\author{
Dong Yin
}
