% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Parameter.r
\name{param.geno}
\alias{param.geno}
\title{Generate parameters for genotype data simulation}
\usage{
param.geno(SP = NULL, ...)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{...}{one or more parameter(s) for genotype simulation}
}
\value{
a list of genotype simulation parameters
}
\description{
Build date: Feb 21, 2022
Last update: Feb 21, 2022
}
\examples{
SP <- param.geno(pop.marker = 1e4, pop.ind = 1e2)
str(SP)
}
\author{
Dong Yin
}
