% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Parameter.r
\name{param.simer}
\alias{param.simer}
\title{Generate parameters for simer}
\usage{
param.simer(SP = NULL, ...)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{...}{one or more parameter(s) for simer}
}
\value{
a list of simer parameters
}
\description{
Build date: Apr 17, 2022
Last update: Apr 17, 2022
}
\examples{
SP <- param.simer(out = "simer")
str(SP)
}
\author{
Dong Yin
}
