% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{paste_label}
\alias{paste_label}
\title{Pasting label}
\usage{
paste_label(line, label, side = "right", margin = 2)
}
\arguments{
\item{line}{long text.}

\item{label}{short label.}

\item{side}{"right" or "left".}

\item{margin}{the margin information, default 2.}
}
\value{
none.
}
\description{
Paste label to a line.
}
\details{
Build date: Oct 22, 2018
Last update: Apr 30, 2022
}
\author{
Dong Yin, Lilin Yin, Haohao Zhang, and Xiaolei Liu
}
\keyword{internal}
