% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Phenotype.r
\name{phenotype}
\alias{phenotype}
\title{Generate genotype and editing genotype}
\usage{
phenotype(SP = NULL, verbose = TRUE)
}
\arguments{
\item{SP}{a list of all simulation parameters}

\item{verbose}{whether to print detail}
}
\value{
a genotype matrix with block and map information
}
\description{
Build date: Nov 14, 2018
Last update: Feb 23, 2022
}
\examples{
pop.env <- list(
  F1 = list(
    level = c("1", "2"),
    eff = list(tr1 = c(50, 30), tr2 = c(50, 30))
  ), 
  F2 = list(
    level = c("d1", "d2", "d3"),
    eff = list(tr1 = c(10, 20, 30), tr2 = c(10, 20, 30))
  ),
  R1 = list(
    level = c("l1", "l2", "l3"),
    ratio = list(tr1 = 0.1, tr2 = 0.1)
  )
)
SP <- param.annot(qtn.num = diag(c(10, 10)), qtn.model = "A + D + A:D")
SP <- param.geno(SP = SP, pop.marker = 1e4, pop.ind = 1e2)
SP <- param.pheno(
  SP = SP, 
  pop.ind = 100,
  pop.rep = 2,
  pop.rep.bal = TRUE,
  pop.env = pop.env,
  phe.var = list(tr1 = 100, tr2 = 100),
  phe.model = list(
    tr1 = "T1 = A + D + A:D + F1 + F2 + R1 + A:F1 + E",
    tr2 = "T2 = A + D + A:D + F1 + F2 + R1 + A:F1 + E"
  )
)
SP <- annotation(SP)
SP <- genotype(SP)
SP <- phenotype(SP)
head(SP$pheno$pop$gen1)
}
\references{
Kao C and Zeng Z (2002) <https://www.genetics.org/content/160/3/1243.long>
}
\author{
Dong Yin
}
