% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selects.r
\name{read.selgeno}
\alias{read.selgeno}
\title{Read sel_ind_geno to get selection condition}
\usage{
read.selgeno(pop = NULL, selPath = NULL, out = NULL)
}
\arguments{
\item{pop}{total population information}

\item{selPath}{the path of breeding_plan}

\item{out}{path of output files}
}
\value{
None
}
\description{
Build date: Nov 14, 2018
Last update: Jul 31, 2019
}
\examples{
\donttest{
# get map file, map is neccessary
data(simdata)

# run simer
simer.list <-
     simer(num.gen = 10,
           replication = 1,
           verbose = TRUE, 
           mrk.dense = FALSE,
           out = NULL,
           out.format = "numeric",
           seed.geno = runif(1, 0, 100),
           seed.map = 12345,
           out.geno.gen = 3:5,
           out.pheno.gen = 1:5,
           rawgeno1 = rawgeno,
           rawgeno2 = NULL,
           rawgeno3 = NULL,
           rawgeno4 = NULL,
           num.ind = NULL,
           prob = c(0.5, 0.5),
           input.map = input.map,
           len.block = 5e7,
           range.hot = 4:6,
           range.cold = 1:5,
           rate.mut = 1e-8,
           cal.model = "A",
           h2.tr1 = 0.3,
           num.qtn.tr1 = 18,
           var.tr1 = 2,
           dist.qtn.tr1 = "normal",
           eff.unit.tr1 = 0.5,
           shape.tr1 = 1,
           scale.tr1 = 1,
           multrait = FALSE,
           num.qtn.trn = matrix(c(18, 10, 10, 20), 2, 2),
           eff.sd = matrix(c(1, 0, 0, 2), 2, 2),
           gnt.cov = matrix(c(1, 2, 2, 15), 2, 2),
           env.cov = matrix(c(10, 5, 5, 100), 2, 2),
           qtn.spot = rep(0.1, 10),
           maf = 0,
           sel.crit = "pheno",
           sel.on = TRUE, 
           mtd.reprod = "randmate",
           userped = userped,
           num.prog = 2,
           ratio = 0.5,
           prog.tri = 2,
           prog.doub = 2,
           prog.back = rep(2, 5),
           ps = 0.8,
           decr = TRUE,
           sel.multi = "index",
           index.wt = c(0.5, 0.5),
           index.tdm = 1,
           goal.perc = 0.1,
           pass.perc = 0.9, 
           sel.sing = "comb") 

pop <- simer.list$pop                         
selPath <- system.file("extdata", "01breeding_plan", package = "simer")
out.pop <- read.selgeno(pop = pop, selPath = selPath, out = NULL)
}
}
\author{
Dong Yin
}
