% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{remove_bigmatrix}
\alias{remove_bigmatrix}
\title{Remove big.matrix safely}
\usage{
remove_bigmatrix(x, desc_suffix = ".geno.desc", bin_suffix = ".geno.bin")
}
\arguments{
\item{x}{filename of big.matrix}

\item{desc_suffix}{suffix of description file of big.matrix}

\item{bin_suffix}{suffix of binary file of big.matrix}
}
\value{
TRUE or FALSE
}
\description{
Build date: Aug 8, 2019
Last update: Feb 14, 2019
}
\examples{
library(bigmemory)
mat <- filebacked.big.matrix(
     nrow = 10,
     ncol = 10,
     init = 0,
     type = 'char',
     backingpath = ".",
     backingfile = 'simer.geno.bin',
     descriptorfile = 'simer.geno.desc')

remove_bigmatrix(x = "simer")
}
\author{
Haohao Zhang and Dong Yin
}
