% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduces.r
\name{reproduces}
\alias{reproduces}
\title{Do reproducion by different mate design}
\usage{
reproduces(pop1 = NULL, pop2 = NULL, pop1.geno = NULL,
  pop2.geno = NULL, ind.stay = NULL, mtd.reprod = "randmate",
  num.prog = 2, ratio = 0.5)
}
\arguments{
\item{pop1}{population information of population1}

\item{pop2}{population information of population2}

\item{pop1.geno}{genotype matrix of population1}

\item{pop2.geno}{genotype matrix of population2}

\item{ind.stay}{selected individuals regarded as parents}

\item{mtd.reprod}{different reproduction methods with the options: "clone", "dh", "selfpol", "singcro", "randmate", and "randexself"}

\item{num.prog}{litter size of dams}

\item{ratio}{ratio of the males in all individuals}
}
\value{
population information and genotype matrix of current population
}
\description{
Build date: Nov 14, 2018
Last update: Aug 1, 2019
}
\examples{
basepop <- getpop(nind = 100, from = 1, ratio = 0.1)
basepop.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)
pop.list <- reproduces(pop1 = basepop,
                      pop1.geno = basepop.geno,
                      ind.stay = basepop$index,
                      mtd.reprod = "randmate",
                      num.prog = 2,
                      ratio = 0.5)
pop <- pop.list$pop
geno <- pop.list$geno
str(pop)
str(geno)
}
\author{
Dong Yin
}
