% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{rule_wrap}
\alias{rule_wrap}
\title{Accomplishment}
\usage{
rule_wrap(string, width, align = "center", linechar = " ")
}
\arguments{
\item{string}{a string.}

\item{width}{the width of the message.}

\item{align}{the position of string.}

\item{linechar}{char in every line.}
}
\value{
none.
}
\description{
Print accomplishment information.
}
\details{
Build date: Oct 22, 2018
Last update: Apr 30, 2022
}
\author{
Dong Yin, Lilin Yin, Haohao Zhang, and Xiaolei Liu
}
\keyword{internal}
