% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selects.r
\name{sel.core}
\alias{sel.core}
\title{Get core population ID and genotype}
\usage{
sel.core(ind.stay = NULL, core.stay = NULL, refresh = rep(0.6, 2),
  keep.max.gen = rep(3, 2), incols = 2, pop.total, pop.geno.curr,
  pop.geno.core)
}
\arguments{
\item{ind.stay}{ID of the sires and the dams passing selection}

\item{core.stay}{ID the core sires and dams}

\item{refresh}{refresh ratio of core population of sires and dams, only used in ps > 1}

\item{keep.max.gen}{if ps <= 1, fraction selected in selection of males and females; if ps > 1, ps is number of selected males and females}

\item{incols}{the column number of an individual in the input genotype matrix, it can be 1 or 2}

\item{pop.total}{total population information}

\item{pop.geno.curr}{genotype matrix of current population}

\item{pop.geno.core}{genotype matrix of core population}
}
\value{
core population list
}
\description{
Build date: May 2, 2020
Last update: May 2, 2020
}
\examples{
pop1 <- getpop(nind = 100, from =   1, ratio = 0.5)
pop2 <- getpop(nind = 100, from = 101, ratio = 0.5, gen = 2)
pop.total <- rbind(pop1, pop2)
incols <- 2
pop.geno.core <- genotype(num.marker=500, num.ind= 80, incols=incols)
pop.geno.curr <- genotype(num.marker=500, num.ind=100, incols=incols)
core.stay <- list(sir =   1: 30, dam =  51:100)
 ind.stay <- list(sir = 101:130, dam = 151:200)
info.core <- sel.core(ind.stay = ind.stay, core.stay = core.stay, incols = incols, 
    refresh = rep(0.6, 2), keep.max.gen = rep(1, 2), pop.total = pop.total, 
    pop.geno.core = pop.geno.core, pop.geno.curr = pop.geno.curr)
str(info.core)

# change the individual 101 to the individual 1
# individual 1 must be removed in the test
core.stay <- info.core$core.stay
pop.geno.core <- info.core$core.geno
core.stay$sir[1] <- 1
pop.geno.core[, 1] <- pop.geno.curr[, 1]

pop3 <- getpop(nind = 100, from = 201, ratio = 0.5, gen = 3)
pop.total <- rbind(pop.total, pop3)
ind.stay <- list(sir = 201:230, dam = 251:300)
pop.geno.curr <- genotype(num.marker=500, num.ind=100, incols=incols)
info.core <- sel.core(ind.stay = ind.stay, core.stay = core.stay, incols = incols, 
    refresh = rep(0.6, 2), keep.max.gen = rep(2, 2), pop.total = pop.total, 
    pop.geno.curr = pop.geno.curr, pop.geno.core = pop.geno.core)
str(info.core)
}
\author{
Dong Yin
}
