% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.r
\name{set.pheno}
\alias{set.pheno}
\title{Set the phenotype of the population}
\usage{
set.pheno(pop, pop.pheno, sel.crit)
}
\arguments{
\item{pop}{population information of generation, family ID, within-family ID, individual ID, paternal ID, maternal ID, and sex}

\item{pop.pheno}{phenotype information}

\item{sel.crit}{selection criteria with the options: "TGV", "TBV", "pEBVs", "gEBVs", "ssEBVs", and "pheno"}
}
\value{
population information with phenotype
}
\description{
Set the phenotype of the population
}
\examples{
pop <- getpop(nind = 100, from = 1, ratio = 0.1)
pop.geno <- genotype(num.marker = 48353, num.ind = 100, verbose = TRUE)
effs <-
    cal.effs(pop.geno = pop.geno,
             cal.model = "A",
             num.qtn.tr1 = c(2, 6, 10),
             sd.tr1 = c(0.4, 0.2, 0.02, 0.02, 0.02, 0.02, 0.02, 0.001),
             dist.qtn.tr1 = rep("normal", 6),
             prob.tr1 = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5),
             shape.tr1 = c(1, 1, 1, 1, 1, 1),
             scale.tr1 = c(1, 1, 1, 1, 1, 1),
             multrait = FALSE,
             num.qtn.trn = matrix(c(18, 10, 10, 20), 2, 2),
             sd.trn = diag(c(1, 0.5)),
             qtn.spot = rep(0.1, 10),
             maf = 0, 
             verbose = TRUE)

pop.pheno <-
    phenotype(effs = effs,
              FR = NULL, 
              cv = list(fam = 0.5), 
              pop = pop,
              pop.geno = pop.geno,
              pos.map = NULL,
              h2.tr1 = c(0.3, 0.1, 0.05, 0.05, 0.05, 0.01),
              gnt.cov = matrix(c(1, 2, 2, 15), 2, 2),
              h2.trn = c(0.3, 0.5),  
              sel.crit = "pheno", 
              pop.total = pop, 
              sel.on = TRUE, 
              inner.env = pop.env, 
              verbose = TRUE)

str(pop)
pop <- set.pheno(pop, pop.pheno, sel.crit = "pheno")
str(pop)
}
