% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Data.R
\name{simer.Data.Env}
\alias{simer.Data.Env}
\title{To find appropriate fixed effects, covariates, and random effects
Author: Dong Yin
Build date: July 17, 2021
Last update: July 17, 2021}
\usage{
simer.Data.Env(
  planPhe,
  fileMVP = NULL,
  filePed = NULL,
  header = TRUE,
  sep = "\\t",
  verbose = TRUE
)
}
\arguments{
\item{planPhe}{breeding plans about phenotype}

\item{fileMVP}{Genotype in MVP format}

\item{filePed}{the pedigree files, it can be a vector}

\item{header}{the header of file}

\item{sep}{the separator of file}

\item{verbose}{whether to print detail.}
}
\value{
the best effects for EBV model
}
\description{
To find appropriate fixed effects, covariates, and random effects
Author: Dong Yin
Build date: July 17, 2021
Last update: July 17, 2021
}
\examples{
jsonFile <- system.file("extdata", "demo2.json", package = "simer")
jsonList <- rjson::fromJSON(file = jsonFile)

# prepare genotype data
genoPath <- jsonList$genotype
genoFiles <- list.files(genoPath)
fileMVP <- grep(pattern = "geno.desc", genoFiles, value = TRUE)
fileMVP <- file.path(genoPath, fileMVP)
fileMVP <- substr(fileMVP, 1, nchar(fileMVP)-10)

# prepare pedigree data
filePed <- jsonList$pedigree

# prepare phenotype data
planPhe <- jsonList$analysis_plan

# reset planPhe by optimized model
# planPhe <- simer.Data.Env(planPhe = planPhe, fileMVP, filePed)
}
