% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.r
\name{simer}
\alias{simer}
\title{Main function of simer}
\usage{
simer(num.gen = 5, replication = 1, verbose = TRUE,
  mrk.dense = FALSE, out = NULL, out.format = "numeric",
  seed.geno = runif(1, 0, 100), seed.map = 12345,
  out.geno.gen = (num.gen - 2):num.gen, out.pheno.gen = 1:num.gen,
  rawgeno1 = NULL, rawgeno2 = NULL, rawgeno3 = NULL,
  rawgeno4 = NULL, num.ind = 100, prob = c(0.5, 0.5),
  input.map = NULL, len.block = 5e+07, range.hot = 4:6,
  range.cold = 1:5, rate.mut = 1e-08, cal.model = "A",
  h2.tr1 = 0.3, num.qtn.tr1 = 18, var.tr1 = 2,
  dist.qtn.tr1 = "normal", eff.unit.tr1 = 0.5, shape.tr1 = 1,
  scale.tr1 = 1, multrait = FALSE, num.qtn.trn = matrix(c(18, 10, 10,
  20), 2, 2), eff.sd = matrix(c(1, 0, 0, 2), 2, 2),
  gnt.cov = matrix(c(14, 10, 10, 15), 2, 2), env.cov = matrix(c(6, 5,
  5, 10), 2, 2), qtn.spot = rep(0.1, 10), maf = 0,
  sel.crit = "pheno", sel.on = TRUE, mtd.reprod = "randmate",
  userped = NULL, num.prog = 2, ratio = 0.5, prog.tri = 2,
  prog.doub = 2, prog.back = rep(2, num.gen), ps = 0.8,
  decr = TRUE, sel.multi = "index", index.wt = c(0.5, 0.5),
  index.tdm = 1, goal.perc = 0.1, pass.perc = 0.9,
  sel.sing = "comb")
}
\arguments{
\item{num.gen}{number of generations in simulation}

\item{replication}{replication index of simulation}

\item{verbose}{whether to print detail}

\item{mrk.dense}{whether markers are dense}

\item{out}{path of output files}

\item{out.format}{format of output, "numeric" or "plink"}

\item{seed.geno}{random seed of genotype matrix}

\item{seed.map}{random seed of map file}

\item{out.geno.gen}{indice of generation of output genotype}

\item{out.pheno.gen}{indice of generation of  output phenotype}

\item{rawgeno1}{extrinsic genotype matrix1}

\item{rawgeno2}{extrinsic genotype matrix2}

\item{rawgeno3}{extrinsic genotype matrix3}

\item{rawgeno4}{extrinsic genotype matrix4}

\item{num.ind}{population size of base population}

\item{prob}{weight of "0" and "1" in genotype matrix, the sum of element in vector equals 1}

\item{input.map}{map from outside}

\item{len.block}{length of every blocks}

\item{range.hot}{range of exchages in hot spot block}

\item{range.cold}{range of exchages in cold spot block}

\item{rate.mut}{mutation rate between 1e-8 and 1e-6}

\item{cal.model}{phenotype model with "A", "AD", "ADI"}

\item{h2.tr1}{heritability vector of trait1, corresponding to a, d, aXa, aXd, dXa, dXd}

\item{num.qtn.tr1}{integer or integer vector, the number of QTN in the trait1}

\item{var.tr1}{variances of different effects, the last 5 vector elements are corrresponding to d, aXa, aXd, dXa, dXd respectively and the rest elements are corresponding to a}

\item{dist.qtn.tr1}{distribution of QTN's effects with options: "normal", "geometry" and "gamma", vector elements are corresponding to a, d, aXa, aXd, dXa, dXd respectively}

\item{eff.unit.tr1}{unit effect of geometric distribution of trait1, vector elements are corresponding to a, d, aXa, aXd, dXa, dXd respectively}

\item{shape.tr1}{shape of gamma distribution of trait1, vector elements are corresponding to a, d, aXa, aXd, dXa, dXd respectively}

\item{scale.tr1}{scale of gamma distribution of trait1, vector elements are corresponding to a, d, aXa, aXd, dXa, dXd respectively}

\item{multrait}{whether applying pair traits with overlapping, TRUE represents applying, FALSE represents not}

\item{num.qtn.trn}{QTN distribution matrix, diagnal elements are total QTN number of the trait, non-diagnal are QTN number of overlop qtn}

\item{eff.sd}{a matrix with the standard deviation of QTN effects}

\item{gnt.cov}{genetic covaiance matrix among all traits}

\item{env.cov}{environment covaiance matrix among all traits}

\item{qtn.spot}{QTN probability in every blocks}

\item{maf}{Minor Allele Frequency, markers selection range is from  maf to 0.5}

\item{sel.crit}{selection criteria with options: "TGV", "TBV", "pEBVs", "gEBVs", "ssEBVs", "pheno"}

\item{sel.on}{whether to add selection}

\item{mtd.reprod}{different reproduction methods with options: "clone", "dh", "selfpol", "singcro", "tricro", "doubcro", "backcro","randmate", "randexself" and "userped"}

\item{userped}{user-specific pedigree}

\item{num.prog}{litter size of dams}

\item{ratio}{ratio of males in all individuals}

\item{prog.tri}{litter size of the first single cross process in trible cross process}

\item{prog.doub}{litter size of the first two single cross process in double cross process}

\item{prog.back}{a vector with litter size in every generations}

\item{ps}{fraction selected in selection}

\item{decr}{whether to sorting with descreasing}

\item{sel.multi}{selection method of multi-trait with options: "tdm", "indcul" and "index"}

\item{index.wt}{economic weights of selection index method}

\item{index.tdm}{index represents which trait is being selected. NOT CONTROL BY USER}

\item{goal.perc}{percentage of goal more than mean of scores of individuals}

\item{pass.perc}{percentage of expected excellent individuals}

\item{sel.sing}{selection method of single trait with options: "ind", "fam", "infam" and "comb"}
}
\value{
a list with population information, genotype matrix, map information, selection intensity
}
\description{
Build date: Jan 7, 2019
Last update: Oct 13, 2019
}
\examples{
\donttest{
# get map file, map is neccessary
data(simdata)

# run simer
simer.list <-
     simer(num.gen = 10,
           replication = 1,
           verbose = TRUE, 
           mrk.dense = FALSE,
           out = NULL,
           out.format = "numeric",
           seed.geno = runif(1, 0, 100),
           seed.map = 12345,
           out.geno.gen = 3:5,
           out.pheno.gen = 1:5,
           rawgeno1 = rawgeno,
           rawgeno2 = NULL,
           rawgeno3 = NULL,
           rawgeno4 = NULL,
           num.ind = NULL,
           prob = c(0.5, 0.5),
           input.map = input.map,
           len.block = 5e7,
           range.hot = 4:6,
           range.cold = 1:5,
           rate.mut = 1e-8,
           cal.model = "A",
           h2.tr1 = 0.3,
           num.qtn.tr1 = 18,
           var.tr1 = 2,
           dist.qtn.tr1 = "normal",
           eff.unit.tr1 = 0.5,
           shape.tr1 = 1,
           scale.tr1 = 1,
           multrait = FALSE,
           num.qtn.trn = matrix(c(18, 10, 10, 20), 2, 2),
           eff.sd = matrix(c(1, 0, 0, 2), 2, 2),
           gnt.cov = matrix(c(1, 2, 2, 15), 2, 2),
           env.cov = matrix(c(10, 5, 5, 100), 2, 2),
           qtn.spot = rep(0.1, 10),
           maf = 0,
           sel.crit = "pheno",
           sel.on = TRUE, 
           mtd.reprod = "randmate",
           userped = userped,
           num.prog = 2,
           ratio = 0.5,
           prog.tri = 2,
           prog.doub = 2,
           prog.back = rep(2, 5),
           ps = 0.8,
           decr = TRUE,
           sel.multi = "index",
           index.wt = c(0.5, 0.5),
           index.tdm = 1,
           goal.perc = 0.1,
           pass.perc = 0.9, 
           sel.sing = "comb") 
pop <- simer.list$pop
effs <- simer.list$effs
trait <- simer.list$trait
geno <- simer.list$geno
map <- simer.list$map
si <- simer.list$si
str(pop)
str(geno[])
str(map)
si
}
}
\author{
Dong Yin, Lilin Yin, Haohao Zhang and Xiaolei Liu
}
