% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{write.file}
\alias{write.file}
\title{Write files of simer}
\usage{
write.file(SP)
}
\arguments{
\item{pop}{population information of generation, family index, within-family index, index, sire, dam, sex, phenotpye}

\item{geno}{genotype matrix of population}

\item{map}{map information of markers}

\item{out.geno.index}{indice of individuals outputting genotype}

\item{out.pheno.index}{indice of individuals outputting phenotype}

\item{out}{prefix of output file name}

\item{outpath}{path of output files}

\item{out.format}{format of output, "numeric" or "plink"}

\item{verbose}{whether to print detail}
}
\value{
None
}
\description{
Build date: Jan 7, 2019
Last update: Oct 16, 2019
}
\examples{
# DO NOT RUN IT
# outpath <- tempdir()
# SP <- param.simer(out = "simer")
# SP <- simer(SP)
# SP$global$outpath <- outpath
# write.file(SP)
# unlink(outpath, recursive = TRUE)
}
\author{
Dong Yin
}
