% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simer.Utility.R
\name{write.file}
\alias{write.file}
\title{File writing}
\usage{
write.file(SP)
}
\arguments{
\item{SP}{a list of all simulation parameters.}
}
\value{
none.
}
\description{
Write files of Simer.
}
\details{
Build date: Jan 7, 2019
Last update: Jan 28, 2025
}
\examples{
\donttest{
outpath <- tempdir()
SP <- param.simer(out = "simer")
SP <- simer(SP)
SP$global$outpath <- outpath
write.file(SP)
unlink(file.path(outpath, "180_Simer_Data_numeric"), recursive = TRUE)
}
}
\author{
Dong Yin
}
